% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustEstimate.R
\name{adjustEstimate}
\alias{adjustEstimate}
\title{Compute Bias Corrected Quantities.}
\usage{
adjustEstimate(epsilon, Omega, dmu, dOmega, n.cluster, name.param,
  name.endogenous, name.meanparam, name.varparam, index.Omega, adjust.Omega,
  adjust.n, tol, n.iter, trace)
}
\description{
Compute bias corrected residuals variance covariance matrix
and information matrix.
Also provides the leverage values and corrected sample size when adjust.n is set to TRUE.
}
\keyword{internal}
