% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntDensTri.R
\name{autoplot.IntDensTri}
\alias{autoplot.IntDensTri}
\title{2D-display of the Domain Used to Compute the Integral}
\usage{
\method{autoplot}{IntDensTri}(object, coord.plot = c("x", "y1"),
  plot = TRUE, ...)
}
\arguments{
\item{object}{output of the function \code{IntDensTri}.}

\item{coord.plot}{[character vector] the x and y coordinates. Can be \code{"x"}, \code{"y1"} to \code{"yd"}, \code{"z"} if \code{zmin} was specified when calling \code{IntDensTri}.}

\item{plot}{[logical] should the plot be displayed?}

\item{...}{[internal] Only used by the generic method.}
}
\value{
A \code{ggplot} object.
}
\description{
2D-display of the domain used to compute the integral.
}
\seealso{
\code{\link{IntDensTri}}
}
