% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistMax.R
\name{calcDistMax}
\alias{calcDistMax}
\title{Adjust the p.values using the quantiles of the max statistic}
\usage{
calcDistMax(statistic, iid, mu, conditional, method, alpha, ncpus, n.sim, trace,
  n.repMax = 100)
}
\arguments{
\item{statistic}{the observed statistic}

\item{iid}{zero-mean iid decomposition of the observed coefficients used to compute the statistic.}

\item{mu}{estimated value for the coefficients}

\item{conditional}{values to condition on. 
If not \code{NULL} the values should correspond the variable in to the first column(s) of the argument iid.}

\item{method}{the method used to compute the p.values. Can be \code{"integration"}, \code{"boot-wild"}, or \code{"boot-norm"}.
See the detail section.}

\item{alpha}{the significance threshold for retaining a new link}

\item{ncpus}{the number of cpu to use for parellel computations}

\item{n.sim}{the total number of simulations.}

\item{trace}{should the execution of the function be traced.}

\item{n.repMax}{the maximum number of rejection when using "\code{"boot-wild"} or \code{"boot-norm"}.}
}
\description{
Adjust the p.values using the quantiles of the max statistic.
}
\examples{
library(mvtnorm)

n <- 100
p <- 5
X.iid <- rmvnorm(n, mean = rep(0,p), sigma = diag(1,p))

calcDistMax(1:p, iid = X.iid, mu = rep(1,p), conditional =  NULL, method = "integration",
            trace = FALSE, alpha = 0.05, ncpus = 1)

calcDistMax(1:p, iid = X.iid, mu = rep(1,p), conditional =  NULL, method = "boot-wild",
            trace = FALSE, alpha = 0.05, ncpus = 1, n.sim = 10)

}
