% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateType1.R
\name{calibrateType}
\alias{calibrateType}
\alias{calibrateType1}
\title{Simulation Study Assessing Bias and Type 1 Error}
\usage{
calibrateType1(object, null, n, n.rep, coef.value, dir.save = NULL,
  n.true = 1e+05, check.true = TRUE, bootstrap = FALSE,
  type.bootstrap = c("perc", "stud", "bca"), n.bootstrap = 1000,
  seed = NULL, trace = 2)
}
\arguments{
\item{object}{a \code{lvm} object defining the generative model.}

\item{null}{[character vector] names of the coefficient whose value will be tested against 0.}

\item{n}{[integer vector, >0] sample size(s) considered in the simulation study.}

\item{n.rep}{[integer, >0] number of simulations per sample size.}

\item{coef.value}{[named numeric vector] values for each coefficient of the model.
Can also be \code{NULL}: in such a case the coefficients are set to default values decided by lava (usually 0 or 1).}

\item{dir.save}{[character] path to the directory were the results should be exported.
Can also be \code{NULL}: in such a case the results are not exported.}

\item{n.true}{[integer, >0] sample size at which the estimated coefficients will be a reliable approximation of the true coefficients.}

\item{check.true}{[logical] should \code{coef.value} be compared to the estimate coefficients when the sample size equals \code{n.true}?}

\item{bootstrap}{[logical] should bootstrap resampling be performed?}

\item{type.bootstrap}{[character vector]}

\item{n.bootstrap}{[integer, >0] the number of bootstrap sample to be used for each bootstrap.}

\item{seed}{[integer, >0] seed value that will be set at the beginning of the simulation to enable eproducibility of the results.
Can also be \code{NULL}: in such a case no seed is set.}

\item{trace}{[interger] should the execution of the function be trace. Can be 0, 1 or 2.}
}
\value{
An object of class \code{calibrateType1}.
}
\description{
Perform a simulation study over one or several sample size
to assess the bias of the estimate
and the type 1 error of the Wald test and robust Wald test
}
\examples{
#### generative model ####
m.Sim <- lvm(c(Y1[mu1:sigma]~1*eta,
               Y2[mu2:sigma]~1*eta,
               Y3[mu3:sigma]~1*eta,
               eta~beta1*Group+beta2*Gender))
latent(m.Sim) <- ~eta
categorical(m.Sim, labels = c("M","F")) <- ~Gender

vec.par <- c(mu2 = 0, ## mu1 is set to 0 by default
             mu3 = -0.37,
             "eta" = -2.74,
             beta1 = -1.20,
             beta2 = 0,
             sigma = 1.46,
             "eta~~eta" = 1.63)

#### parameters to test ####
null <- c("mu2","beta2")

#### launch simulation ####
\dontrun{
res <- calibrateType1(m.Sim, null = null, n = c(20,30,40), n.rep = 50, coef.value = vec.par)
autoplot(res, type = "bias")
autoplot(res, type = "type1error")
}
\dontshow{
res <- calibrateType1(m.Sim, null = null, n = c(20,30,40), n.rep = 2, coef.value = vec.par)
}
}
\seealso{
\code{link{autoplot.calibrateType1}} for a graphical display of the bias or of the type 1 error.
}
\author{
Brice Ozenne
}
