% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm2.R
\name{coef-multcomp}
\alias{coef-multcomp}
\alias{coef.mmm2}
\alias{coef.ls.lvmfit}
\title{Model Coefficients}
\usage{
\method{coef}{mmm2}(object, ...)

\method{coef}{ls.lvmfit}(object, ...)
}
\arguments{
\item{object}{a \code{mmm2} or \code{ls.lvmfit} object.}

\item{...}{Not used. Only for compatibility with the generic method.}
}
\description{
Returns the model coefficients of a \code{mmm2} or \code{ls.lvmfit} object.
For internal use.
}
\keyword{internal}
