% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-compareSearch.R
\name{compareSearch}
\alias{compareSearch}
\title{Compare methods to identify missing local dependencies in a LVM}
\usage{
compareSearch(object, alpha = 0.05, method.iid, method.p.adjust, statistic,
  trace = 1, ...)
}
\arguments{
\item{object}{a lvm model}

\item{alpha}{the significance threshold for retaining a new link.}

\item{method.iid}{the method used to compute the influence function. Can be \code{"iid"} (first order approximation) or \code{"iidJack"} (jacknife estimate).}

\item{method.p.adjust}{the method used to adjust the p.values for multiple comparisons. Ignore when using the max statistic.}

\item{statistic}{test statistic used to perform the test. Can the likelihood ratio test (LR), the score (score) or the max statistic (max).}

\item{trace}{should the execution be traced?}

\item{...}{additional arguments to be passed to lower levels functions.}
}
\value{
a latent variable model
}
\description{
Compare methods to identify missing local dependencies in a LVM
}
\examples{
mSim <- lvm(Y~G+X1+X2)
addvar(mSim) <- ~Z1+Z2+Z3+Z4+Z5+Z6
dt <- as.data.table(lava::sim(mSim, 1e2))

mBase <- lvm(Y~G)
addvar(mBase) <- ~X1+X2+Z1+Z2+Z3+Z4+Z5+Z6
e.lvm <- estimate(mBase, data = dt)

res <- modelsearch2(e.lvm, statistic = "Wald", method.p.adjust = "none")
res$sequenceTest
\dontrun{
res <- compareSearch(e.lvm, statistic = c("score","Wald"),
                     method.iid = "iid",
                     method.p.adjust = c("holm","fdr","max"))
}
res

}
