% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-conditionalMoment.R
\name{conditionalMoment}
\alias{conditionalMoment}
\alias{conditionalMoment.lm}
\alias{conditionalMoment.gls}
\alias{conditionalMoment.lme}
\alias{conditionalMoment.lvm}
\alias{conditionalMoment.lvmfit}
\title{Prepare the Computation of score2}
\usage{
conditionalMoment(object, initialize, first.order, second.order, usefit,
  residuals, leverage, param = NULL, data = NULL)

\method{conditionalMoment}{lm}(object, initialize, first.order,
  second.order, usefit, residuals, leverage, param = NULL, data = NULL)

\method{conditionalMoment}{gls}(object, initialize, first.order,
  second.order, usefit, residuals, leverage, param = NULL, data = NULL)

\method{conditionalMoment}{lme}(object, initialize, first.order,
  second.order, usefit, residuals, leverage, param = NULL, data = NULL)

\method{conditionalMoment}{lvm}(object, initialize, first.order,
  second.order, usefit, residuals, leverage, param = NULL, data = NULL)

\method{conditionalMoment}{lvmfit}(object, initialize, first.order,
  second.order, usefit, residuals, leverage, param = NULL, data = NULL)
}
\arguments{
\item{object, x}{a latent variable model.}

\item{first.order}{[logical] should the terms relative to the first derivative of the conditional moments be considered?}

\item{second.order}{[logical] should the terms relative to the second derivative of the conditional moments be considered?}

\item{usefit}{[logical] If TRUE the coefficients estimated by the model are used to pre-compute quantities.}

\item{param}{[numeric vector] the fitted coefficients.}

\item{data}{[data.frame] data set.}

\item{cluster}{[list] information about the cluster and endogenous variables (output of getGroups2).}

\item{...}{[internal] only used by the generic method or by the <- methods.}
}
\description{
Compute the conditional mean and variance,
and their first and second derivative regarding the model parameters.
}
\details{
For lvmfit objects, there are two levels of pre-computation:
\itemize{
\item a basic one that do no involve the model coefficient (\code{conditionalMoment.lvm}).
\item an advanced one that require the model coefficients (\code{conditionalMoment.lvmfit}). 
}
}
\examples{
m <- lvm(Y1~eta,Y2~eta,Y3~eta)
latent(m) <- ~eta

d <- lava::sim(m,1e2)
e <- estimate(m, d)

## basic pre-computation
res1 <- conditionalMoment(e, data = d,
                         first.order = FALSE, second.order = FALSE,
                         name.endogenous = endogenous(e),
                         name.latent = latent(e), usefit = FALSE)
res1$skeleton$Sigma

## full pre-computation
res2 <- conditionalMoment(e, param = coef(e), data = d,
                         first.order = FALSE, second.order = FALSE,
                         name.endogenous = endogenous(e),
                         name.latent = latent(e), usefit = TRUE
)
res2$value$Sigma

}
\concept{derivative of the score equation}
\concept{small sample inference}
\keyword{internal}
