% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContrast.R
\name{contrast2name}
\alias{contrast2name}
\alias{.contrast2name}
\title{Create Rownames for a Contrast Matrix}
\usage{
.contrast2name(C, null = NULL)
}
\arguments{
\item{C}{a constrast matrix.}
}
\description{
Create rownames for a contrast matrix.
using the coefficients and the names of the parameters
}
