% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContrast.R
\name{createContrast}
\alias{createContrast}
\alias{createContrast.character}
\alias{createContrast.lm}
\alias{createContrast.gls}
\alias{createContrast.lme}
\alias{createContrast.lvmfit}
\alias{createContrast.list}
\title{Contrast matrix for multiple latent variable models}
\usage{
createContrast(object, ...)

\method{createContrast}{character}(object, name.param, add.rowname = TRUE,
  ...)

\method{createContrast}{lm}(object, par, ...)

\method{createContrast}{gls}(object, par, ...)

\method{createContrast}{lme}(object, par, ...)

\method{createContrast}{lvmfit}(object, par, ...)

\method{createContrast}{list}(object, par = NULL, var.test = NULL, ...)
}
\arguments{
\item{object}{an \code{ls.lvmfit} object.}

\item{...}{Only used by the generic method.}

\item{name.param}{[internal] the names of all the model parameters.}

\item{add.rowname}{[internal] should a name be defined for each hypothesis.}

\item{par}{[vector of characters] expression defining the linear hypotheses to be tested. See the examples section.}

\item{var.test}{[character] a regular expression that is used to identify the coefficients to be tested using \code{grep}. Each coefficient will be tested in a separate hypothesis. When this argument is used, the argument \code{par} is disregarded.}
}
\description{
Returns an empty contrast matrix corresponding to a list of latent variable models.
}
\examples{
## Simulate data
mSim <- lvm(X ~ Age + Treatment,
            Y ~ Gender + Treatment,
            c(Z1,Z2,Z3) ~ eta, eta ~ treatment,
            Age[40:5]~1)
latent(mSim) <- ~eta
categorical(mSim, labels = c("placebo","SSRI")) <- ~Treatment
categorical(mSim, labels = c("male","female")) <- ~Gender
n <- 1e2
set.seed(10)
df.data <- sim(mSim,n)

## Estimate separate models
lmX <- estimate(lvm(X ~ -1 + Age + Treatment), data = df.data)
lmY <- estimate(lvm(Y ~ -1 + Gender + Treatment), data = df.data)
lvmZ <- estimate(lvm(c(Z1,Z2,Z3) ~ -1 + 1*eta, eta ~ -1 + Treatment), 
                 data = df.data)

## Contrast matrix for a given model
createContrast(lmX, "X~Age")
createContrast(lmX, c("X~Age=0","X~Age+5*X~TreatmentSSRI=0"))

## Contrast matrix for the join model
ls.lvm <- list(X = lmX, Y = lmY, Z = lvmZ)
createContrast(ls.lvm, var.test = "Treatment")

}
