% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntDensTri.R
\name{createGrid}
\alias{createGrid}
\title{Create a mesh for the integration}
\usage{
createGrid(n, xmin, xmax, plot = FALSE)
}
\arguments{
\item{n}{the number of points for the mesh in one direction.}

\item{xmin}{the minimal x value.}

\item{xmax}{the maximal x value.}

\item{plot}{should the mesh be displayed}
}
\description{
Create a mesh for the integration
}
\examples{
dt.res <- lavaSearch2:::createGrid(20, xmin = 0, xmax = 4)
}

