% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCategoricalLink.R
\name{defineCategoricalLink}
\alias{defineCategoricalLink}
\alias{defineCategoricalLink.lvm}
\title{Identify categorical links in lvm}
\usage{
defineCategoricalLink(x, ...)

\method{defineCategoricalLink}{lvm}(x, link = NULL, data = NULL)
}
\arguments{
\item{x}{a lvm model.}

\item{link}{the links to be analyzed. If NULL, all the coefficients from the lvm model are used instead.}

\item{data}{the dataset that will be used to fit the model. If \code{NULL}, a simulated data will be generated from the model.}

\item{Y.name}{the outcome variable of each link.}

\item{X.name}{the regression variable of each link.}
}
\description{
Identify categorical links in lvm
}
\examples{
 
m <- lvm(Y1~X1+X2,Y2~X1+X3)
categorical(m, K = 3) <- "X1"
try(defineCategoricalLink(m)) # error

categorical(m, K = 3, labels = 1:3) <- "X1"
defineCategoricalLink(m)
defineCategoricalLink(m, "Y~X1")
defineCategoricalLink(m, "X1:0|1")
defineCategoricalLink(m, "X1:1|2")
defineCategoricalLink(m, c("X1:0|1", "X1:1|2"))
defineCategoricalLink(m, c("Y~X1","Y~X2"))
defineCategoricalLink(m, c("Y~X2","Y~X1"))

}
