% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeOfFreedom.R
\name{degreeOfFreedom}
\alias{degreeOfFreedom}
\alias{degreeOfFreedom.lvmfit}
\title{Approximate the degree of freedoms of a lvm model}
\usage{
degreeOfFreedom(x, ...)

\method{degreeOfFreedom}{lvmfit}(x, conservative, ...)
}
\arguments{
\item{x}{model object}

\item{...}{additional arguments}

\item{conservative}{If true the total number of parameter is substracted from the number of observations. If false only the number of mean parameters is substracted.}
}
\description{
Approximate the degree of freedoms of a lvm model
}
\examples{
n <- 20

set.seed(10)
m <- lvm(Y~X1+X2+X3)
d <- sim(m, n)
e <- estimate(m, data = d)
degreeOfFreedom(e, conservative = FALSE)
degreeOfFreedom(e, conservative = TRUE)

}
