% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.residual.R
\name{df.residual}
\alias{df.residual}
\alias{df.residual.lvmfit}
\alias{df.residual.coxph}
\title{Approximate degree of freedoms of a model}
\usage{
\method{df.residual}{lvmfit}(object, power = 1/2, adjust.residuals = TRUE,
  as.clubSandwich = TRUE, dmu.dtheta = NULL, vcov.param,
  ls.leverage.adj = NULL, Omega = NULL, iOmega = NULL,
  Omega_chol = NULL, ...)

\method{df.residual}{coxph}(object, ...)
}
\arguments{
\item{object}{model object}

\item{power}{the type of leverage-adjusted residuals to be used. See the details section.}

\item{adjust.residuals}{should leverage-adjusted residuals be used?}

\item{Omega}{}

\item{iOmega}{}

\item{...}{additional arguments}

\item{dmu.dtheta2}{}

\item{leverage.adj}{}

\item{data}{}

\item{p}{}
}
\description{
Approximate the degree of freedoms of a model
}
\details{
Leverage-adjusted residuals have been shown to improve the coverage of robust standard errors in small samples.

Since the calculation of the degrees of freedom involves the use of the residuals,
the same small sample correction should be use to compute the degrees of freedom and the robust standard errors.
}
\examples{
n <- 20

set.seed(10)
m <- lvm(Y~X1+X2+X3+X4)
d <- lava::sim(m, n)
e <- estimate(m, data = d)
df.residual(e)


}
