% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2.R
\name{dfSigmaRobust}
\alias{dfSigmaRobust}
\title{Degree of Freedom for the Robust Chi-Square Test}
\usage{
dfSigmaRobust(contrast, vcov, rvcov, score)
}
\arguments{
\item{contrast}{[numeric vector] the linear combination of parameters to test}

\item{vcov}{[numeric matrix] the variance-covariance matrix of the parameters.}

\item{rvcov}{[numeric matrix] the robust variance-covariance matrix of the parameters.}

\item{score}{[numeric matrix] the individual score for each parameter.}
}
\description{
Computation of the degrees of freedom of the chi-squared distribution
relative to the robust-based variance
}
\details{
When contrast is the identity matrix, this function compute the moments of the sandwich estimator
and the degrees of freedom of the approximate t-test as described in (Pan, 2002) section 2 and 3.1.
}
\references{
Wei Pan and Melanie M. Wall, Small-sample adjustments in using the sandwich variance estiamtor in generalized estimating equations. Statistics in medicine (2002) 21:1429-1441.
}
