% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfVariance.R
\name{dfVariance}
\alias{dfVariance}
\alias{dfVariance.gls}
\alias{dfVariance.lme}
\alias{dfVariance.lvmfit}
\title{Compute the degree of freedom of the variance parameters}
\usage{
dfVariance(object, ...)

\method{dfVariance}{gls}(object, cluster, robust = FALSE,
  adjust.residuals = FALSE, power = 0.5, as.clubSandwich = TRUE, ...)

\method{dfVariance}{lme}(object, cluster, robust = FALSE,
  adjust.residuals = FALSE, power = 0.5, as.clubSandwich = TRUE, ...)

\method{dfVariance}{lvmfit}(object, fix.mean = FALSE, robust = FALSE,
  adjust.residuals = FALSE, power = 0.5, as.clubSandwich = TRUE, ...)
}
\arguments{
\item{object}{a lvm object}

\item{cluster}{the grouping variable relative to which the observations are i.i.d.}
}
\description{
Compute the degree of freedom of the variance parameters
}
