% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-dInformation2.R
\name{.dinformation2-internal}
\alias{.dinformation2-internal}
\alias{.dInformation2}
\title{Compute the First Derivative of the Expected Information Matrix}
\usage{
.dInformation2(
  dmu,
  dOmega,
  d2mu,
  d2Omega,
  OmegaM1,
  missing.pattern,
  unique.pattern,
  name.pattern,
  grid.3varD1,
  grid.2meanD1.1varD1,
  grid.2meanD2.1meanD1,
  grid.2varD2.1varD1,
  name.param,
  leverage,
  n.cluster,
  weights
)
}
\description{
Compute the first derivative of the expected information matrix.
}
\details{
\code{calc_dinformation} will perform the computation individually when the
argument \code{index.Omega} is not null.
}
\keyword{internal}
