% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate2.R
\name{estimate2}
\alias{estimate2}
\alias{estimate2.lm}
\alias{estimate2.gls}
\alias{estimate2.lme}
\alias{estimate2.lvmfit}
\alias{estimate2.lvmfit2}
\alias{estimate2<-}
\alias{estimate2<-.lm}
\alias{estimate2<-.gls}
\alias{estimate2<-.lme}
\alias{estimate2<-.lvmfit}
\alias{estimate2<-.lvmfit2}
\title{Compute the Derivative of the Information Matrix}
\usage{
estimate2(object, ...)

\method{estimate2}{lm}(object, adjust.residuals = FALSE,
  return.score = FALSE, ...)

\method{estimate2}{gls}(object, cluster, vcov.param = NULL,
  adjust.residuals = FALSE, numericDerivative = FALSE,
  return.score = FALSE, ...)

\method{estimate2}{lme}(object, cluster, vcov.param = NULL,
  adjust.residuals = FALSE, numericDerivative = FALSE,
  return.score = FALSE, ...)

\method{estimate2}{lvmfit}(object, vcov.param = NULL,
  adjust.residuals = TRUE, numericDerivative = FALSE,
  return.score = FALSE, ...)

\method{estimate2}{lvmfit2}(object, ...)

estimate2(x, ...) <- value

\method{estimate2}{lm}(x, ...) <- value

\method{estimate2}{gls}(x, ...) <- value

\method{estimate2}{lme}(x, ...) <- value

\method{estimate2}{lvmfit}(x, ...) <- value

\method{estimate2}{lvmfit2}(x, ...) <- value
}
\arguments{
\item{object}{a gls, lme, or lvm object.}

\item{...}{[internal] Only used by the generic method.}

\item{adjust.residuals}{[logical] small sample correction:
should the leverage-adjusted residuals be used to compute the score?
Otherwise the raw residuals will be used.}

\item{return.score}{[for internal use] export the score.}

\item{cluster}{the grouping variable relative to which the observations are iid.
Only required for gls models with no correlation argument.}

\item{vcov.param}{the variance-covariance matrix of the estimates.}

\item{numericDerivative}{If TRUE, the degree of freedom are computed using a numerical derivative.}

\item{x}{same as object.}

\item{value}{same as adjust.residuals.}
}
\description{
Compute the derivative of the information matrix.
}
