% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-butils-evalInParentEnv.R
\name{evalInParentEnv}
\alias{evalInParentEnv}
\title{Find Object in the Parent Environments}
\usage{
evalInParentEnv(name, envir)
}
\arguments{
\item{name}{[character] the name of the object to get.}

\item{envir}{[environments] the environment from which to look for the object.}
}
\description{
Search an object in the parent environments. For internal use.
}
\concept{
extractor
}
\keyword{internal}
