% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\title{Extract data from a model}
\usage{
extractData(object, coxExpand = FALSE, force = FALSE, convert2dt = TRUE)
}
\arguments{
\item{object}{the fitted model.}

\item{coxExpand}{should the dataset be exported as a start/stop?
Otherwise it will be exported as a Surv column}

\item{force}{if \code{TRUE} and if the first attempt to extract the data failed, 
then the data are extracted using the name of the dataset.}

\item{convert2dt}{should the object containing the data be converted into a data.table?}
}
\description{
Extract data from a model using \code{nlme::getData}, \code{riskRegression::CoxDesign} or \code{model.frame}.. 
If it fails it will try to extract it by its name according to \code{model$call$data}.
}
\examples{
set.seed(10)
n <- 100

#### linear regression ####
Y1 <- rnorm(n, mean = 0)
Y2 <- rnorm(n, mean = 0.3)
df <- rbind(data.frame(Y=Y1,G=1,Id = 1:5),
           data.frame(Y=Y2,G=2,Id = 1:5)
           )
m.lm <- lm(Y ~ G, data = df)
extractData(m.lm)

library(nlme)
m.gls <- gls(Y ~ G, weights = varIdent(form = ~ 1|Id), data = df)
extractData(m.gls)
m.lme <- lme(Y ~ G, random = ~ 1|Id, data = df)
extractData(m.lme)

library(lava)
e <- estimate(lvm(Y ~ G), data = df)
extractData(e)

#### survival ####
library(riskRegression)
library(survival)
dt.surv <- sampleData(n, outcome = "survival")
m.cox <- coxph(Surv(time, event) ~ X1 + X2, data = dt.surv, x = TRUE, y = TRUE)
extractData(m.cox)
m.cox <- coxph(Surv(time, event) ~ strata(X1) + X2, data = dt.surv, x = TRUE, y = TRUE)
xx <- extractData(m.cox)

#### nested fuuctions ####
fct1 <- function(m){
   fct2(m)
}
fct2 <- function(m){ 
   extractData(m)
}
fct1(m.gls)
}
