% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitplot.R
\name{fitplot}
\alias{fitplot}
\title{Plot the fit of a lvm object}
\usage{
fitplot(object, ...)
}
\arguments{
\item{object}{a lvm model.}

\item{...}{additional arguments.}

\item{variables}{the endogenous variable for which the fit should be displayed.}

\item{plot}{should the fit be displayed in a graphical window.}
}
\value{
a list containing:
\itemize{
\item plot the display of the fit
\item data the dataset used to make the display
}
}
\description{
Plot the fit for each outcome of a lvm object
}
\examples{
m <- lvm(y1~eta+x1,y2~eta,y3~eta+x3)
latent(m) <- ~ eta
dd <- sim(m,100) ## Simulate 100 observations from model
e <- estimate(m, dd) ## Estimate parameters

fitplot(e)

}

