% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-nlme.R
\name{getCluster2-internal}
\alias{getCluster2-internal}
\alias{.getCluster2}
\alias{.getCluster2.gls}
\alias{.getCluster2.lme}
\title{Reconstruct the Cluster Variable from a nlme Model}
\usage{
.getCluster2(object, ...)

\method{.getCluster2}{gls}(object, cluster, data, ...)

\method{.getCluster2}{lme}(object, ...)
}
\arguments{
\item{object}{a \code{gls} or \code{lme} object.}

\item{...}{[internal] Only used by the generic method.}

\item{cluster}{[integer vector] the grouping variable relative to which the observations are iid.
Only required for \code{gls} models with no correlation argument.}

\item{data}{[data.frame] the data set.}
}
\value{
A list containing:
\itemize{
\item cluster: the cluster index for each observation.
\item n.cluster: the number of clusters.
}
}
\description{
Reconstruct the cluster variable from a nlme model.
}
\concept{extractor}
\keyword{internal}
