% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-getGroups2.R
\name{getGroup2-internal}
\alias{getGroup2-internal}
\alias{.getGroups2}
\title{Reconstruct the Cluster variable}
\usage{
.getGroups2(object, data, index.Omega, endogenous)
}
\arguments{
\item{object}{a \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} object.}

\item{data}{dataset.}

\item{cluster}{[integer vector] the grouping variable relative to which the observations are iid.
Only required for \code{gls} models with no correlation argument.}

\item{...}{[internal] Only used by the generic method.}
}
\value{
A list containing:
\itemize{
\item index.cluster: the cluster index for each observation.
\item name.cluster: a unique identifier for each cluster.
\item n.cluster: the number of clusters.
}
}
\description{
Reconstruct the cluster variable.
Similar to \code{nlme::getGroups}.
}
\concept{extractor}
\keyword{internal}
