% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-nlme.R
\name{getGroup2}
\alias{getGroup2}
\alias{.getGroups2}
\alias{.getGroups2.gls}
\alias{getGroup2}
\alias{.getGroups2.lme}
\title{Reconstruct the Cluster Variable from a nlme Model}
\usage{
.getGroups2(object, ...)

\method{.getGroups2}{gls}(object, cluster, data, ...)

\method{.getGroups2}{lme}(object, ...)
}
\arguments{
\item{object}{a \code{gls} or \code{lme} object.}

\item{...}{[internal] Only used by the generic method.}

\item{cluster}{[vector] the grouping variable relative to which the observations are iid.
Only required for \code{gls} models with no correlation argument.}

\item{data}{[data.frame] the data set.}
}
\value{
A list containing:
\itemize{
\item cluster: the cluster index for each observation.
\item n.cluster: the number of clusters.
\item endogenous: to which endogenous variable each observation corresponds to.
\item n.endogenous: the number of endogenous variables.
\item index.obs: a vector to convert observations from the vector format to the matrix format.
}
}
\description{
Reconstruct the cluster variable from a nlme model.
}
\details{
The variance coefficients that are exported are the residual variance of each outcome.
This is \eqn{\sigma^2} for the first one and \eqn{k^2 \sigma^2} for the remaining ones.
}
\concept{
extractor
}
\keyword{internal}
