% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-getIndexOmega.R
\name{getIndexOmega-internal}
\alias{getIndexOmega-internal}
\alias{.getIndexOmega}
\alias{.getIndexOmega.lm}
\alias{.getIndexOmega.gls}
\alias{.getIndexOmega.lme}
\alias{.getIndexOmega.lvmfit}
\title{Identify the Endogenous Variables}
\usage{
.getIndexOmega(object, data, ...)

\method{.getIndexOmega}{lm}(object, data, ...)

\method{.getIndexOmega}{gls}(object, data, ...)

\method{.getIndexOmega}{lme}(object, data, ...)

\method{.getIndexOmega}{lvmfit}(object, data, ...)
}
\arguments{
\item{object}{a \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} object.}

\item{data}{dataset.}

\item{...}{[internal] Only used by the generic method.}
}
\description{
Identify the endogenous variables, i.e., returns a vector with length the number of observations,
whose values are the index of the repetitions.
}
\examples{
#### simulate data ####
set.seed(10)
dW <- sampleRepeated(10, format = "wide")
set.seed(10)
dL <- sampleRepeated(10, format = "long")
dL$time2 <- paste0("visit",dL$time)

#### linear model ####
e.lm <- lm(Y1~X1, data = dW)
.getIndexOmega(e.lm, data = dW)

#### gls model ####
e.gls1 <- gls(Y1~X1, data = dW)
.getIndexOmega(e.gls1, data = dW)

e.gls2 <- gls(Y~X1, correlation = corCompSymm(form=~1|id), data = dL)
.getIndexOmega(e.gls2, data = dL)

e.gls3 <- gls(Y~X1, correlation = corCompSymm(form=~time|id), data = dL)
.getIndexOmega(e.gls3, data = dL)

e.gls4 <- gls(Y~X1, weight = varIdent(form=~1|time2), data = dL)
.getIndexOmega(e.gls4, data = dL)

e.gls5 <- gls(Y~X1, weight = varIdent(form=~1|time2),
              correlation = corSymm(form=~time|id), data = dL)
.getIndexOmega(e.gls5, data = dL)

#### lme model ####
e.lme <- lme(Y~X1, random=~1|id, data = dL)
.getIndexOmega(e.lme, data = dL)

#### lvm model ####
e.lvm <- estimate(lvm(c(Y1,Y2,Y3) ~ 1*eta + X1, eta ~ Z1), data = dW)
.getIndexOmega(e.lvm, data = dW)

}
\concept{extractor}
\keyword{internal}
