% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInParentEnv.R
\name{getInParentEnv}
\alias{getInParentEnv}
\title{Find object in the parent environments}
\usage{
getInParentEnv(name, envir)
}
\arguments{
\item{name}{character string containing the name of the object to get.}

\item{envir}{the environment from which to look for the object.}
}
\description{
Internal function
}
