% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-mlf2.R
\name{glht.mlf2}
\alias{glht.mlf2}
\title{General linear hypothesis}
\usage{
\method{glht}{mlf2}(model, linfct, ...)
}
\arguments{
\item{model}{a mmm2 object}

\item{linfct}{a specification of the linear hypotheses to be tested.
See the documentation of the glht function from the multcomp package.}
}
\description{
General linear hypotheses and multiple outcomes for mmm2 objects.
}
