% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-hessian2.R
\name{hessian2-internal}
\alias{hessian2-internal}
\alias{.hessian2}
\title{Compute the Hessian Matrix From the Conditional Moments}
\usage{
.hessian2(
  dmu,
  dOmega,
  d2mu,
  d2Omega,
  epsilon,
  OmegaM1,
  missing.pattern,
  unique.pattern,
  name.pattern,
  grid.mean,
  grid.var,
  grid.hybrid,
  name.param,
  leverage,
  n.cluster,
  weights
)
}
\description{
Compute the Hessian matrix from the conditional moments.
}
\details{
\code{calc_hessian} will perform the computation individually when the
argument \code{index.Omega} is not null.
}
\keyword{internal}
