% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-hessian2.R
\name{hessian2}
\alias{hessian2}
\alias{hessian2.sCorrect}
\title{Extract the Hessian After Small Sample Correction.}
\usage{
hessian2(object, cluster, as.lava)

\method{hessian2}{sCorrect}(object, cluster = NULL, as.lava = TRUE)
}
\arguments{
\item{object}{a \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} object.}

\item{cluster}{[integer vector] the grouping variable relative to which the observations are iid.}

\item{as.lava}{[logical] Should the order and the name of the coefficient be the same as those obtained using coef with type = -1.
Only relevant for \code{lvmfit} objects.}
}
\value{
An array containing the second derivative of the likelihood relative to each sample (dim 3)
and each pair of model coefficients (dim 1,2).
}
\description{
Extract the hessian from \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} objects.
}
\examples{
n <- 5e1
p <- 3
X.name <- paste0("X",1:p)
link.lvm <- paste0("Y~",X.name)
formula.lvm <- as.formula(paste0("Y~",paste0(X.name,collapse="+")))

m <- lvm(formula.lvm)
distribution(m,~Id) <- sequence.lvm(0)
set.seed(10)
d <- lava::sim(m,n)

## linear model
e.lm <- lm(formula.lvm,data=d)
eS.lm <- sCorrect(e.lm, ssc = NA, df = "Satterthwaite")
hessian2.tempo <- hessian2(eS.lm)
colMeans(score.tempo)

## latent variable model
e.lvm <- estimate(lvm(formula.lvm),data=d)
eS.lvm <- sCorrect(e.lvm, df = NA, ssc = NA)
score.tempo <- score2(eS.lvm, indiv = TRUE)
range(score.tempo-score(e.lvm, indiv = TRUE))

}
\seealso{
\code{\link{sCorrect}} to obtain \code{lm2}, \code{gls2}, \code{lme2}, or \code{lvmfit2} objects.
}
\concept{small sample inference}
