% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook.R
\name{hook.lavaSearch2}
\alias{hook.lavaSearch2}
\alias{gethook_lavaSearch2}
\alias{addhook_lavaSearch2}
\title{Hooks for lavaSearch2}
\usage{
gethook_lavaSearch2(hook, ...)

addhook_lavaSearch2(x, hook, ...)
}
\arguments{
\item{hook}{the name of the hook}

\item{...}{for compatibility with lava}

\item{x}{the function to add to the hook}
}
\description{
Get and add hook for lavaSeach2
}
