% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-iid2.R
\name{iid2}
\alias{iid2}
\alias{iid2.lm}
\alias{iid2.gls}
\alias{iid2.lme}
\alias{iid2.lvmfit}
\title{Extract i.i.d. decomposition from linear and latent variable models}
\usage{
iid2(x, ...)

\method{iid2}{lm}(x, data = NULL, adjust.residuals = TRUE, power = 1/2,
  ...)

\method{iid2}{gls}(x, p = NULL, data = NULL, adjust.residuals = TRUE,
  power = 1/2, ...)

\method{iid2}{lme}(x, p = NULL, data = NULL, adjust.residuals = TRUE,
  power = 1/2, ...)

\method{iid2}{lvmfit}(x, p = NULL, data = NULL, adjust.residuals = TRUE,
  power = 1/2, as.clubSandwich = TRUE, ...)
}
\arguments{
\item{x}{a linear model or a latent variable model}

\item{...}{arguments to be passed to \code{score2}.}

\item{data}{[optional] data set.}

\item{adjust.residuals}{Should the leverage-adjusted residuals be used to compute the score? Otherwise the raw residuals will be used.#'}

\item{power}{the exponent used for computing the leverage-adjusted residuals. See the details section.}

\item{indiv}{Should the score relative to each observation be exported? Otherwise the total score (i.e. sum over all observations) will be exported.#'}
}
\value{
A matrix.
}
\description{
Extract i.i.d. decomposition from linear and latent variable models using leverage adjusted fitted residuals.
}
\details{
Leverage-adjusted residuals have been shown to improve the coverage of robust standard errors in small samples.
They are computed according to the formula:
\eqn{e_adj = \frac{e}{(1-h_{ii})^\alpha}}

\code{power = 0} (i.e. code{adjust.residuals = FALSE}) corresponds to the use of the raw residuals.
It can be shown that the variance-covariance estimator is biased downward for the mean parameters in LVM.

\code{power=0.5} corrects for the bias of the variance-covariance estimator,
at least for the mean parameters (Kauermann and Carroll, 2001)..
This corresponds to the correction "CR2" in the clubSandwich package.

\code{power=1} approximates the variance-covariance jackknife estimator (Bell and McCaffrey, 2002).
This corresponds to the correction "CR3" in the clubSandwich package.
}
\examples{
n <- 5e1
p <- 3
X.name <- paste0("X",1:p)
link.lvm <- paste0("Y~",X.name)
formula.lvm <- as.formula(paste0("Y~",paste0(X.name,collapse="+")))

m <- lvm(formula.lvm)
distribution(m,~Id) <- sequence.lvm(0)
set.seed(10)
d <- sim(m,n)

e.lm <- lm(formula.lvm,data=d)
iid2(e.lm)
iid2(e.lm,corrected=FALSE)-iid(e.lm)

e.lvm <- estimate(lvm(formula.lvm),data=d)
score(e.lvm)
range(iid2(e.lvm, adjust.residuals = FALSE)-iid(e.lvm))

#### multiple regression
m <- lvm(c(Y1~X1,Y2~X2,Y3~X3))
e.lvm <- estimate(m,data=sim(m,1e2))
iid2(e.lvm)

}
\references{
Bell, R. M., & McCaffrey, D. F. Bias reduction in standard errors for linear regression with multi-stage samples. Survey Methodology, 28(2), 169-181 (2002). \cr
Kauermann G. and Carroll R. J. A note on the efficiency of sandwich covariance matrix estimation. Journal of the American Statistical Association. Vol. 96, No. 456 (2001).
}
