% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-information2.R
\name{information2-internal}
\alias{information2-internal}
\alias{.information2}
\title{Compute the Expected Information Matrix From the Conditional Moments}
\usage{
.information2(
  dmu,
  dOmega,
  OmegaM1,
  missing.pattern,
  unique.pattern,
  name.pattern,
  grid.mean,
  grid.var,
  name.param,
  leverage,
  weights = NULL,
  n.cluster
)
}
\description{
Compute the expected information matrix from the conditional moments.
}
\details{
\code{calc_information} will perform the computation individually when the
argument \code{index.Omega} is not null.
}
\keyword{internal}
