% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-information2.R
\name{information2}
\alias{information2}
\alias{information2.sCorrect}
\title{Extract The Full Information Matrix After Small Sample Correction.}
\usage{
information2(object, as.lava)

\method{information2}{sCorrect}(object, as.lava = TRUE)
}
\arguments{
\item{object}{a \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} object.}

\item{as.lava}{[logical] Should the order and the name of the coefficient be the same as those obtained using coef with type = -1.
Only relevant for \code{lvmfit} objects.}
}
\value{
A matrix.
}
\description{
Extract the full information matrix from  from \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} objects.
Similar to \code{lava::information} but with small sample correction.
}
\examples{
n <- 5e1
p <- 3
X.name <- paste0("X",1:p)
link.lvm <- paste0("Y~",X.name)
formula.lvm <- as.formula(paste0("Y~",paste0(X.name,collapse="+")))

m <- lvm(formula.lvm)
distribution(m,~Id) <- sequence.lvm(0)
set.seed(10)
d <- lava::sim(m,n)

## linear model
e.lm <- lm(formula.lvm,data=d)
info.tempo <- vcov2(e.lm)
info.tempo[names(coef(e.lm)),names(coef(e.lm))] - vcov(e.lm)

## latent variable model
e.lvm <- estimate(lvm(formula.lvm),data=d)
vcov.tempo <- vcov2(e.lvm)
round(vcov.tempo \%*\% information(e.lvm), 5)

}
\seealso{
\code{\link{sCorrect}} to obtain \code{lm2}, \code{gls2}, \code{lme2}, or \code{lvmfit2} objects.
}
\concept{small sample inference}
