% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initVar_link.R
\name{initVar}
\alias{initVar}
\alias{initVar_link}
\alias{initVar_links}
\title{Normalize var1 and var2}
\usage{
initVar_link(var1, var2, repVar1 = FALSE, format = "list",
  Slink = lava.options()$symbols[1], Scov = lava.options()$symbols[2])

initVar_links(var1, format = "list")
}
\arguments{
\item{var1}{a character indicating the endogeneous variable or a formula}

\item{var2}{an optional character indicating the exogeneous variable}

\item{repVar1}{should var1 be duplicated to match var2 length. Only active if format = "list".}

\item{format}{should the name of the variable be return (format = "list"), a vector of character formula ("txt.formula") or a list of formula ("formula")}

\item{Slink}{the symbol for regression link}

\item{Scov}{the symbol for covariance link}
}
\description{
Convert var1 and var2 from formula or covariance to character
}
\details{
See test file test/testthat/test-Reduce.R for examples
}
\examples{
initVar_link(y ~ x1)
initVar_link("y ~ x1")
initVar_link(y ~ x1 + x2)
initVar_link("y ~ x1 + x2")
initVar_link(y ~ x1 + x2, repVar1 = TRUE)
initVar_link(y ~ x1 + x2, repVar1 = TRUE, format = "formula")
initVar_link(y ~ x1 + x2, repVar1 = TRUE, format = "txt.formula")
initVar_link("y", "x1", format = "formula")

initVar_link("y ~ x1:0|1")

initVar_links(y ~ x1)
initVar_links("y ~ x1")
initVar_links(c("y ~ x1","y~ x2"))
initVar_links(c(y ~ x1,y ~ x2))
initVar_links(c("y ~ x1","y~ x2"), format = "formula")
initVar_links(c(y ~ x1,y ~ x2), format = "formula")
initVar_links(c("y ~ x1","y~ x2"), format = "txt.formula")
initVar_links(c(y ~ x1,y ~ x2), format = "txt.formula")
}
