% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntDensTri.R
\name{IntDensTri}
\alias{IntDensTri}
\title{Integration of a density over a triangle}
\usage{
IntDensTri(mu, Sigma, df, n, xmin, xmax, distribution = "pmvnorm")
}
\arguments{
\item{mu}{the mean vector}

\item{Sigma}{the variance-covariance matrix}

\item{df}{degree of freedom for a student distribution.}

\item{n}{number of points in one direction for the mesh}

\item{xmin}{the minimum value along the x axis}

\item{xmax}{the maximum value along the x axis}

\item{distribution}{Can be \code{"pmvnorm"} (normal distribution) or \code{"pvmt"} (student's t distribution)}
}
\description{
Integration of a density over a triangle
}
\examples{
library(mvtnorm)

p <- 2
Sigma <- diag(p)
mu <- rep(0, p)

## bivariate normal distribution
z2 <- qmvt(0.975, mean = mu, sigma = Sigma, df = 1e3)$quantile

# compute integral
IntDensTri(mu = mu, Sigma = Sigma, n=5, xmin=0, xmax=3)-1/2
IntDensTri(mu = mu, Sigma = Sigma, n=5, xmin=0, xmax=5)-1/2
IntDensTri(mu = mu, Sigma = Sigma, n=5, xmin=0, xmax=100)-1/2

IntDensTri(mu = mu, Sigma = Sigma, df = 5, n=5, xmin=0, xmax=3, distribution = "pmvt")-1/2
IntDensTri(mu = mu, Sigma = Sigma, df = 5, n=5, xmin=0, xmax=5, distribution = "pmvt")-1/2
IntDensTri(mu = mu, Sigma = Sigma, df = 5, n=5, xmin=0, xmax=100, distribution = "pmvt")-1/2

## trivariate normal distribution
p <- 3
Sigma <- diag(p)
mu <- rep(0, p)

IntDensTri(mu = mu, Sigma = Sigma, n=5, xmin=c(0,0), xmax=c(10,10))-1/2

}
\author{
Brice Ozenne
}

