% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaSearch2-package.R
\docType{package}
\name{lavaSearch}
\alias{lavaSearch}
\alias{lavaSearch-package}
\title{Tools for model specification in the latent variable framework}
\description{
Tools for model specification in the latent variable framework (add-on for the lava package).
}
