% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-leverage2.R
\name{leverage2}
\alias{leverage2}
\alias{leverage2.sCorrect}
\title{Extract Leverage Values}
\usage{
leverage2(object, format)

\method{leverage2}{sCorrect}(object, format = "wide")
}
\arguments{
\item{object}{a \code{lm2}, \code{gls2}, \code{lme2}, or \code{lvmfit2} object.}

\item{format}{[character] Use \code{"wide"} to return the residuals in the wide format (one row relative to each sample).
Otherwise use \code{"long"} to return the residuals in the long format.}
}
\value{
a matrix containing the leverage relative to each sample (in rows)
and each endogenous variable (in column).
}
\description{
Extract leverage values from a Gaussian linear model.
}
\details{
The leverage are defined as the partial derivative of the fitted values with respect to the observations.
\deqn{
leverage_i = \frac{\partial \hat{Y}_i}{\partial Y_i}
}
See Wei et al. (1998). \cr \cr

If argument \code{p} or \code{data} is not null, then the small sample size correction is recomputed to correct the residuals.
}
\examples{
## simulate data
set.seed(10)
m <- lvm(Y1~eta,Y2~eta,Y3~eta)
latent(m) <- ~eta
d <- lava::sim(m,20, latent = FALSE)

## standard linear model
e.lm <- lm(Y1~Y2, data = d)

sCorrect(e.lm) <- TRUE
range(as.double(leverage2(e.lm)) - influence(e.lm)$hat)

## latent variable model
e.lvm <- estimate(m, data = d)
sCorrect(e.lvm) <- TRUE
leverage2(e.lvm)

}
\references{
Bo-Cheng Wei et al., Generalized Leverage and its applications (1998), Scandinavian Journal of Statistics 25:1:25-37.
}
\seealso{
\code{\link{sCorrect}} to obtain \code{lm2}, \code{gls2}, \code{lme2}, or \code{lvmfit2} objects.
}
\concept{small sample inference}
