% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-lTest.R
\name{lTest}
\alias{lTest}
\alias{lTest.lm}
\alias{lTest.gls}
\alias{lTest.lme}
\alias{lTest.lvmfit}
\title{Compute the degree of freedom of the variance parameters}
\usage{
lTest(object, ...)

\method{lTest}{lm}(object, C = NULL, adjust.residuals = TRUE,
  Ftest = TRUE, ...)

\method{lTest}{gls}(object, C = NULL, adjust.residuals = TRUE,
  Ftest = TRUE, ...)

\method{lTest}{lme}(object, C = NULL, adjust.residuals = TRUE,
  Ftest = TRUE, ...)

\method{lTest}{lvmfit}(object, C = NULL, adjust.residuals = TRUE,
  Ftest = TRUE, ...)
}
\arguments{
\item{object}{a lvm object}

\item{cluster}{the grouping variable relative to which the observations are i.i.d.}
}
\description{
Compute the degree of freedom of the variance parameters
}
\examples{
set.seed(10)
mSim <- lvm(Y~0.1*X1+0.2*X2)
categorical(mSim, labels = c("a","b","c")) <- ~X1
transform(mSim, Id~Y) <- function(x){1:NROW(x)}
df.data <- sim(mSim, 1e2)

## gold standard
anova(e.lm)

e.lm <- lm(Y~X1+X2, data = df.data)
lTest(e.lm)

## gls model
e.gls <- gls(Y~X1+X2, data = df.data, method = "ML")
e.gls$dVcov <- dVcov2(e.gls, cluster = df.data$Id)

C <- rbind(c(0,1,0,0,0),c(0,0,1,0,0))
colnames(C) <- names(attr(e.gls$dVcov,"param"))
lTest(e.gls, C = C)

C <- rbind(c(0,0,0,1,0))
colnames(C) <- names(attr(e.gls$dVcov,"param"))
lTest(e.gls, C = C)

## latent variable model
m <- lvm(Y~X1+X2)
e.lvm <- estimate(m, df.data)
e.lvm$dVcov <- dVcov2(e.lvm)

C <- rbind(c(0,0,1,0,0),c(0,0,0,1,0))
colnames(C) <- names(coef(e.lvm))
lTest(e.lvm, C = C)

C <- rbind(c(0,1,0,0,0))
colnames(C) <- names(coef(e.lvm))
lTest(e.lvm, C = C)

}
