% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlf2.R
\name{mlf2}
\alias{mlf2}
\title{Simultaneous Inference for Multiple Models}
\usage{
mlf2(...)
}
\arguments{
\item{...}{A names argument list containing the fitted model.
See the documentation of \code{multcomp::mmm} for more details.}
}
\description{
Calculation of correlation between test statistics from multiple marginal models using the score decomposition. Similar to \code{multcomp::mmm} but with small sample corrections.
}
