% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsearchLR.R
\name{modelsearchLR}
\alias{modelsearchLR}
\title{Model searching using a likelihood ratio test}
\usage{
modelsearchLR(x, data = NULL, restricted, link, directive,
  ls.LVMargs = NULL, method.p.adjust = "bonferroni",
  display.warnings = TRUE, trace = 1)
}
\value{
a list
}
\description{
Model searching using a likelihood ratio test
}
\keyword{internal}
