% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsearchMax.R
\name{modelsearchMax}
\alias{modelsearchMax}
\title{Model searching using the max statistic}
\usage{
modelsearchMax(x, restricted, link, directive, packages, update.FCT,
  update.args, iid.FCT, method.p.adjust, method.max = "integration",
  n.sim = 1000, alpha = 0.05, iid.previous = NULL,
  quantile.previous = NULL, export.iid = 1, trace = 1, ncpus = 1,
  initCpus = TRUE)
}
\value{
an object of class lvmfit
}
\description{
Model searching using the max statistic to retain or not a link
}
\seealso{
\code{link{modelsearch2}}
}
\keyword{internal}
