% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareScore2.R
\name{prepareScore2}
\alias{prepareScore2}
\alias{prepareScore2.gls}
\alias{prepareScore2.lme}
\alias{prepareScore2<-}
\alias{prepareScore2<-.lvmfit}
\alias{prepareScore2.lvm}
\alias{prepareScore2.lvmfit}
\title{Prepare the computation of score2.}
\usage{
prepareScore2(object, ...)

\method{prepareScore2}{gls}(object, X, param, attr.param, second.order,
  n.cluster, n.endogenous, name.endogenous, index.obs)

\method{prepareScore2}{lme}(object, X, param, attr.param, second.order,
  n.cluster, n.endogenous, name.endogenous, index.obs)

prepareScore2(object, ...) <- value

\method{prepareScore2}{lvmfit}(object, ...) <- value

\method{prepareScore2}{lvm}(object, data, name.endogenous = NULL,
  name.latent = NULL)

\method{prepareScore2}{lvmfit}(object, data = NULL, p = NULL,
  name.endogenous = NULL, name.latent = NULL, second.order = FALSE)
}
\arguments{
\item{object}{a latent variable model}

\item{name.endogenous}{[optional] name of the endogenous variables}

\item{data}{[optional] data set.}

\item{name.latent}{[optional] name of the latent variables}
}
\description{
Compute partial derivatives regarding to the mean and the variance, and compute the design matrices.
}
\examples{
m <- lvm(Y1~eta,Y2~eta,Y3~eta)
latent(m) <- ~eta

e <- estimate(m,sim(m,1e2))
res <- prepareScore2(e)
res$skeleton$dt.param
}
