% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{residplot}
\alias{residplot}
\alias{residplot.lvmfit}
\title{Plot the residuals of a lvm object}
\usage{
residplot(object, ...)

\method{residplot}{lvmfit}(object, variables = NULL, mfrow = NULL,
  type = "qqtest", centralPercents = 0.95, ...)
}
\arguments{
\item{object}{a lvm model.}

\item{...}{additional arguments to be passed to qqtest.}

\item{variables}{the variable for which the residuals should be displayed.}

\item{mfrow}{how to divide the window. See \code{mfrow}.}

\item{type}{the function used to display the qqplot. Can be qqtest or qqnorm.}

\item{centralPercents}{argument passed to \code{qqtest}. See the help of \code{\link{qqtest}}.}
}
\value{
a data frame/cvlvm object containing the convergence status (by default 0 indicates successful convergence, see ?optim), the value of the log-likelihood and the estimated parameters (in columns) for each initialization (in rows)
}
\description{
Plot the residuals of a lvm object
}
\details{
Simulation is based on a multivariate truncated normal law (even though it is not satifying for the variance components)
}
\examples{
m <- lvm(list(y~v1+v2+v3+v4,c(v1,v2,v3,v4)~x))
covariance(m) <- v1~v2+v3+v4
latent(m) <- ~ x
dd <- sim(m,100) ## Simulate 100 observations from model
e <- estimate(m, dd) ## Estimate parameters

residplot(e)

}
