% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-residuals2.R
\name{residuals2}
\alias{residuals2}
\alias{residuals2.gls}
\alias{residuals2.lme}
\alias{residuals2.lvmfit}
\title{Compute the residuals from a lvmfit objects}
\usage{
residuals2(object, ...)

\method{residuals2}{gls}(object, cluster = NULL, p = NULL, data = NULL,
  adjust.residuals = TRUE, power = 1/2, as.clubSandwich = TRUE,
  second.order = FALSE, return.vcov.param = FALSE,
  return.prepareScore2 = FALSE, ...)

\method{residuals2}{lme}(object, cluster = NULL, p = NULL, data = NULL,
  adjust.residuals = TRUE, power = 1/2, as.clubSandwich = TRUE,
  second.order = FALSE, return.vcov.param = FALSE,
  return.prepareScore2 = FALSE, ...)

\method{residuals2}{lvmfit}(object, p = NULL, data = NULL,
  adjust.residuals = TRUE, power = 1/2, as.clubSandwich = TRUE,
  second.order = FALSE, return.vcov.param = FALSE,
  return.prepareScore2 = FALSE, ...)
}
\arguments{
\item{p}{[optional] vector of parameters at which to evaluate the score.}

\item{data}{[optional] data set.}

\item{x}{a fitted latent variable model.}
}
\description{
Compute the residuals from a lvmfit objects
}
\examples{
m <- lvm(Y1~eta,Y2~eta,Y3~eta)
latent(m) <- ~eta

e <- estimate(m,sim(m,1e2))
residuals2(e)
}
