% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-score2.R
\name{score2-internal}
\alias{score2-internal}
\alias{.score2}
\title{Compute the Corrected Score.}
\usage{
.score2(
  dmu,
  dOmega,
  epsilon,
  OmegaM1,
  missing.pattern,
  unique.pattern,
  name.pattern,
  name.param,
  name.meanparam,
  name.varparam,
  n.cluster,
  weights
)
}
\arguments{
\item{n.cluster}{[integer >0] the number of observations.}
}
\description{
Compute the corrected score.
}
\keyword{internal}
