% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-score2.R
\name{score2}
\alias{score2}
\alias{score2.gls}
\alias{score2.lme}
\alias{score2.lvmfit}
\title{Compute the score directly from the gaussian density}
\usage{
score2(x, ...)

\method{score2}{gls}(object, cluster, p = NULL, data = NULL,
  adjust.residuals = TRUE, power = 1/2, indiv = TRUE,
  as.clubSandwich = TRUE, return.vcov.param = FALSE, ...)

\method{score2}{lme}(object, cluster, p = NULL, data = NULL,
  adjust.residuals = TRUE, power = 1/2, indiv = TRUE,
  as.clubSandwich = TRUE, return.vcov.param = FALSE, ...)

\method{score2}{lvmfit}(object, p = NULL, data = NULL,
  adjust.residuals = TRUE, power = 1/2, as.clubSandwich = TRUE,
  indiv = TRUE, return.vcov.param = FALSE, ...)
}
\arguments{
\item{x}{a linear model or a latent variable model.}

\item{...}{not used.}

\item{p}{[optional] vector of parameters at which to evaluate the score.}

\item{data}{[optional] data set.}

\item{adjust.residuals}{Should the leverage-adjusted residuals be used to compute the score? Otherwise the raw residuals will be used.}

\item{power}{the exponent used for computing the leverage-adjusted residuals. See the documentation of the \code{\link{iid2}} function for more details.}

\item{indiv}{Should the score relative to each observation be exported? Otherwise the total score (i.e. sum over all observations) will be exported.}

\item{as.clubSandwich}{The method use to apply the \code{power} argument.}
}
\value{
A matrix.
}
\description{
Compute the score directly from the gaussian density
}
\details{
The log-likelihood of a lvm can written:
\deqn{
  l(\theta|Y,X) = \sum_{i=1}^{n} - \frac{p}{2} log(2\pi) - \frac{1}{2} log|\Omega(\theta))| - \frac{1}{2} (Y_i-\mu_i(\theta)) \Omega^{-1} (Y_i-\mu_i(\theta))^t
}
So the score is:
\deqn{
  s(\theta|Y,X) = \sum_{i=1}^{n}
- \frac{1}{2} tr(\Omega(\theta)^{-1} \frac{\partial \Omega(\theta)}{\partial \theta})
+  \frac{\partial \mu_i(\theta)}{\partial \theta} \Omega^{-1} (Y_i-\mu_i(\theta))^t 
+ \frac{1}{2} (Y_i-\mu_i(\theta)) \Omega^{-1} \frac{\partial \Omega(\theta)}{\partial \theta} \Omega^{-1} (Y_i-\mu_i(\theta))^t
}
with:
\deqn{ \frac{\partial \mu_i(\beta)}{\partial \nu} = 1 }
\deqn{ \frac{\partial \mu_i(\beta)}{\partial K} = X_i }
\deqn{ \frac{\partial \mu_i(\beta)}{\partial \alpha} = (1-B)^{-1}\Lambda }
\deqn{ \frac{\partial \mu_i(\beta)}{\partial \Gamma} = X_i(1-B)^{-1}\Lambda }
\deqn{ \frac{\partial \mu_i(\beta)}{\partial \lambda} = (\alpha + X_i \Gamma)(1-B)^{-1}\delta_{\lambda \in \Lambda} }
\deqn{ \frac{\partial \mu_i(\beta)}{\partial b} = (\alpha + X_i \Gamma)(1-B)^{-1}\delta_{b \in B}(1-B)^{-1}\Lambda }
and:
\deqn{ \frac{\partial \Omega(\beta)}{\partial \psi} = \Lambda^t (1-B)^{-t} \delta_{\psi \in \Psi} (1-B) \Lambda }
\deqn{ \frac{\partial \Omega(\beta)}{\partial \sigma} = \delta_{\sigma \in \Sigma} }
\deqn{ \frac{\partial \Omega(\beta)}{\partial \lambda} = \delta_{\lambda \in \Lambda} (1-B)^{-t} \Psi (1-B)^{-1} \Lambda + \Lambda^t (1-B)^{-t} \Psi (1-B)^{-1} \delta_{\lambda \in \Lambda} }
\deqn{ \frac{\partial \Omega(\beta)}{\partial b} = \Lambda^t (1-B)^{-t} \delta_{b \in B} (1-B)^{-t} \Psi (1-B)^{-1} \Lambda - \Lambda^t (1-B)^{-t} \Psi (1-B)^{-1} \delta_{b \in B} (1-B)^{-1} \Lambda}
}
\examples{

m <- lvm(Y1~eta,Y2~eta,Y3~eta)
latent(m) <- ~eta

e <- estimate(m,sim(m,1e2))
score2(e)

}
