% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initVar_link.R
\name{select.regressor}
\alias{select.regressor}
\alias{select.regressor.formula}
\title{Regressor of a formula}
\usage{
select.regressor(x, ...)

\method{select.regressor}{formula}(x, type = "call")
}
\arguments{
\item{x}{a formula}

\item{...}{additional arguments to be passed to lower levels functions.}

\item{type}{either return an object of type call (\code{"call"}) or the names of the variables (\code{"vars"})}
}
\description{
Return the regressor variables contained in the formula
}
\examples{
select.regressor(Y1~X1+X2)
select.regressor(Y1~X1+X2, type = "vars")

select.regressor(Y1~X1+Y1)
select.regressor(Y1+Y2~X1+Y1, type = "vars")

select.regressor(~X1+X2)
select.regressor(~X1+X2, type = "vars")
}
