% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initVar_link.R
\name{select.response}
\alias{select.response}
\alias{select.response.formula}
\title{Response variable of a formula}
\usage{
select.response(x, ...)

\method{select.response}{formula}(x, type = "call")
}
\arguments{
\item{x}{a formula}

\item{...}{additional arguments to be passed to lower levels functions.}

\item{type}{either return an object of type call (\code{"call"}) or the names of the variables (\code{"vars"})}
}
\description{
Return the reponse variable contained in the formula
}
\examples{
\dontrun{
# select.response <- lavaSearch2:::select.response
select.response(Y1~X1+X2)
select.response(Y1~X1+X2, type = "vars")
select.response(Surv(event,time)~X1+X2, type = "vars")

select.response(Y1~X1+Y1)
select.response(Y1+Y2~X1+Y1, type = "vars")

select.response(~X1+X2)
select.response(~X1+X2, type = "vars")
}
}

