% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initVarlink.R
\name{selectRegressor}
\alias{selectRegressor}
\alias{selectRegressor.formula}
\title{Regressor of a Formula.}
\usage{
selectRegressor(x, ...)

\method{selectRegressor}{formula}(x, type = "call", ...)
}
\arguments{
\item{x}{a formula}

\item{...}{additional arguments to be passed to lower levels functions.}

\item{type}{either return an object of type call (\code{"call"}) or the names of the variables (\code{"vars"})}
}
\description{
Return the regressor variables contained in the formula
}
\examples{

selectRegressor(Y1~X1+X2)
selectRegressor(Y1~X1+X2, type = "vars")

selectRegressor(Y1~X1+Y1)
selectRegressor(Y1+Y2~X1+Y1, type = "vars")

selectRegressor(~X1+X2)
selectRegressor(~X1+X2, type = "vars")

}
