% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initVarlink.R
\name{selectResponse}
\alias{selectResponse}
\alias{selectResponse.formula}
\title{Response Variable of a Formula}
\usage{
selectResponse(x, ...)

\method{selectResponse}{formula}(x, type = "call", ...)
}
\arguments{
\item{x}{a formula}

\item{...}{additional arguments to be passed to lower levels functions.}

\item{type}{either return an object of type call (\code{"call"}) or the names of the variables (\code{"vars"})}
}
\description{
Return the reponse variable contained in the formula.
}
\examples{
selectResponse(Y1~X1+X2)
selectResponse(Y1~X1+X2, type = "vars")
selectResponse(Surv(event,time)~X1+X2, type = "vars")

selectResponse(Y1~X1+Y1)
selectResponse(Y1+Y2~X1+Y1, type = "vars")

selectResponse(~X1+X2)
selectResponse(~X1+X2, type = "vars")

}
