% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvCheck.R
\name{summary.cvlvm}
\alias{summary.cvlvm}
\title{Summary function associated with cvCheck}
\usage{
\method{summary}{cvlvm}(object, threshold = NULL, ...)
}
\arguments{
\item{object}{the output of cvCheck}

\item{threshold}{threshold used to cluster the convergence points (height in hclust)}

\item{...}{for compatibility only}
}
\description{
Summary function associated with cvCheck
}
