% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.glht2.R
\name{summary.glht2}
\alias{summary.glht2}
\title{summary Method for glht2 Objects}
\usage{
\method{summary}{glht2}(object, print = TRUE, ...)
}
\arguments{
\item{object}{output of the \code{glht2} function.}

\item{print}{should the summary be printed in the terminal.}

\item{...}{arguments to be passed to \code{summary.glht}.}
}
\description{
summary method for glht2 objects.
}
