% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modelsearch2.R
\name{summary.modelsearch2}
\alias{summary.modelsearch2}
\title{summary Method for modelsearch2 Objects}
\usage{
\method{summary}{modelsearch2}(object, display = TRUE, ...)
}
\description{
summary method for modelsearch2 objects.
}
