% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-summary2.R
\name{summary2}
\alias{summary2}
\alias{summary2.lm}
\alias{summary2.gls}
\alias{summary2.lme}
\alias{summary2.lvmfit}
\alias{summary2.sCorrect}
\title{Summary After Small Sample Correction}
\usage{
summary2(object, ...)

\method{summary2}{lm}(object, ssc = lava.options()$ssc,
  df = lava.options()$df, ...)

\method{summary2}{gls}(object, ssc = lava.options()$ssc,
  df = lava.options()$df, ...)

\method{summary2}{lme}(object, ssc = lava.options()$ssc,
  df = lava.options()$df, ...)

\method{summary2}{lvmfit}(object, ssc = lava.options()$ssc,
  df = lava.options()$df, ...)

\method{summary2}{sCorrect}(object, robust = FALSE, digit = max(3,
  getOption("digit")), ...)
}
\arguments{
\item{object}{a \code{lm}, \code{gls}, \code{lme} or \code{lvm} object.}

\item{...}{[logical] argument passed to lower level methods.}

\item{ssc}{[logical] should the standard errors of the coefficients be corrected for small sample bias?}

\item{df}{[logical] should the degree of freedoms of the Wald statistic be computed using the Satterthwaite correction?}

\item{robust}{[logical] should the robust standard errors be used instead of the model based standard errors?}

\item{digit}{[integer > 0] the number of decimal places to use when displaying the summary.}

\item{cluster}{[integer vector] the grouping variable relative to which the observations are iid.}
}
\description{
Summary of the estimates, standard errors, and p-values based on Wald tests for \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} objects.
Similar to \code{summary} but with small sample correction (if any).
}
\details{
\code{summary2} is the same as \code{summary}
except that it first computes the small sample correction (but does not store it).
So if \code{summary2} is to be called several times,
it is more efficient to pre-compute the quantities for the small sample correction
using \code{sCorrect} and then call \code{summary2}.

\code{summary2} returns an object with an element \code{table2} containing the estimates, standard errors, degrees of freedom,
upper and lower limits of the confidence intervals, test statistics, and p-values.
}
\examples{
m <- lvm(Y~X1+X2)
set.seed(10)
d <- lava::sim(m, 2e1)

## Gold standard
summary(lm(Y~X1+X2, d))$coef

## gls models
library(nlme)
e.gls <- gls(Y~X1+X2, data = d, method = "ML")
summary(e.gls)$tTable
sCorrect(e.gls, cluster = 1:NROW(d)) <- FALSE ## no small sample correction
summary2(e.gls)$tTable

sCorrect(e.gls, cluster = 1:NROW(d)) <- TRUE ## small sample correction
summary2(e.gls)$tTable

## lvm models
e.lvm <- estimate(m, data = d)
summary(e.lvm)$coef

sCorrect(e.lvm) <- FALSE ## no small sample correction
summary2(e.lvm)$coef

sCorrect(e.lvm) <- TRUE ## small sample correction
summary2(e.lvm)$coef

}
\seealso{
\code{\link{sCorrect}} for more detail about the small sample correction.
}
\concept{small sample inference}
