% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm2.R
\name{vcov-multcomp}
\alias{vcov-multcomp}
\alias{vcov.mmm2}
\alias{vcov.ls.lvmfit}
\title{Variance-Covariance Matrix for a Fitted Object}
\usage{
\method{vcov}{mmm2}(object, return.null = TRUE, adjust.residuals = TRUE,
  robust = FALSE, ...)

\method{vcov}{ls.lvmfit}(object, return.null = TRUE,
  adjust.residuals = TRUE, robust = FALSE, ...)
}
\arguments{
\item{object}{a \code{mmm2} or \code{ls.lvmfit} object.}

\item{return.null}{if TRUE return a matrix filled with NA.}

\item{adjust.residuals}{small sample adjustement.}

\item{robust}{should robust standard error be used? Otherwise rescale the influence function with the standard error obtained from the information matrix.}

\item{...}{Not used. Only for compatibility with the generic method.}
}
\description{
Returns the variance-covariance matrix of a fitted object.
For internal use.
}
