% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-moments2.R
\name{vcov2-internal}
\alias{vcov2-internal}
\alias{.info2vcov}
\title{Inverse the Information Matrix}
\usage{
.info2vcov(information, attr.info = FALSE)
}
\arguments{
\item{attr.info}{[logical] should the information matrix be returned as an attribute?}

\item{...}{arguments passed to .information2}
}
\description{
Compute the inverse of the information matrix.
}
\keyword{internal}
