% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-vcov2.R
\name{vcov2}
\alias{vcov2}
\alias{vcov2.sCorrect}
\title{Extract the Variance Covariance Matrix of the Model Parameters.}
\usage{
vcov2(object, as.lava)

\method{vcov2}{sCorrect}(object, as.lava = TRUE)
}
\arguments{
\item{object}{a \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} object.}

\item{as.lava}{[logical] Should the order and the name of the coefficient be the same as those obtained using coef with type = -1.
Only relevant for \code{lvmfit} objects.}
}
\value{
A matrix.
}
\description{
Extract the variance covariance matrix of the model parameters from \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} objects.
Similar to \code{vcov} but compatible with small sample correction.
}
\details{
If argument \code{p} or \code{data} is not null, then the small sample size correction is recomputed to correct the influence function.
}
\examples{
n <- 5e1
p <- 3
X.name <- paste0("X",1:p)
link.lvm <- paste0("Y~",X.name)
formula.lvm <- as.formula(paste0("Y~",paste0(X.name,collapse="+")))

m <- lvm(formula.lvm)
distribution(m,~Id) <- sequence.lvm(0)
set.seed(10)
d <- lava::sim(m,n)

## linear model
e.lm <- lm(formula.lvm,data=d)
vcov.tempo <- vcov2(e.lm)
vcov.tempo[rownames(vcov(e.lm)),colnames(vcov(e.lm))]/vcov(e.lm)

## latent variable model
e.lvm <- estimate(lvm(formula.lvm),data=d)
vcov.tempo <- vcov2(e.lvm)
vcov.tempo/vcov(e.lvm)

}
\seealso{
\code{\link{sCorrect}} to obtain \code{lm2}, \code{gls2}, \code{lme2}, or \code{lvmfit2} objects.
}
\concept{small sample inference}
