% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{combination}
\alias{combination}
\alias{.combination}
\title{Form all Unique Combinations Between two Vectors}
\usage{
.combination(...)
}
\arguments{
\item{...}{[vectors] elements to be combined.}
}
\value{
A matrix, each row being a different combination.
}
\description{
Form all unique combinations between two vectors (removing symmetric combinations).
}
\examples{
.combination <- lavaSearch2:::.combination

.combination(1,1)
.combination(1:2,1:2)
.combination(c(1:2,1:2),1:2)

.combination(alpha = 1:2, beta = 3:4)

}
\keyword{internal}
