% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-lTest.R
\name{contrast2name}
\alias{contrast2name}
\alias{.contrast2name}
\title{Create rownames for a contrast matrix}
\usage{
.contrast2name(C)
}
\arguments{
\item{C}{a constrast matrix.}
}
\description{
Create rownames for a contrast matrix
using the coefficients and the names of the parameters
}
