% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeOfFreedom.R
\name{df.residual}
\alias{df.residual}
\alias{df.residual.lvmfit}
\alias{df.residual.coxph}
\title{Approximate degree of freedoms of a model}
\usage{
\method{df.residual}{lvmfit}(object, conservative, ...)

\method{df.residual}{coxph}(object, ...)
}
\arguments{
\item{object}{model object}

\item{conservative}{If true the total number of parameter is substracted from the number of observations. If false only the number of mean parameters is substracted.}

\item{...}{additional arguments}
}
\description{
Approximate the degree of freedoms of a model
}
\examples{
n <- 1000

set.seed(10)
m <- lvm(Y~X1+X2+X3+X4)
d <- sim(m, n)
e <- estimate(m, data = d)
df.residual(e, conservative = FALSE)
df.residual(e, conservative = TRUE)


}
