% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2.R
\name{dfSigma}
\alias{dfSigma}
\title{Degree of Freedom for the Chi-Square Test}
\usage{
dfSigma(contrast, vcov, dVcov, keep.param)
}
\arguments{
\item{contrast}{[numeric vector] the linear combination of parameters to test}

\item{vcov}{[numeric matrix] the variance-covariance matrix of the parameters.}

\item{dVcov}{[numeric array] the first derivative of the variance-covariance matrix of the parameters.}

\item{keep.param}{[character vector] the name of the parameters with non-zero first derivative of their variance parameter.}
}
\description{
Computation of the degrees of freedom of the chi-squared distribution
relative to the model-based variance
}
