% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects2.R
\name{effects2}
\alias{effects2}
\alias{effects2.lvmfit2}
\title{Effects from a fitted model}
\usage{
effects2(object, link, ...)

\method{effects2}{lvmfit2}(object, link)
}
\arguments{
\item{object}{an object that inherits from lvmfit.}

\item{link}{[character vector] The path for which the effect should be assessed (e.g. \code{"A~B"}),
i.e. the effect of the right variable (B) on the left variable (A).}

\item{...}{Arguments passed to \code{sCorrect} for \code{lvmfit} objects.
Ignored for \code{lvmfit2} objects.}
}
\description{
Test whether a path in the latent variable model correspond to a null effect.
Similar to \code{lava::effects} but with small sample correction.
So far it only work for paths composed of two edges.
}
\concept{small sample inference}
