% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-effects2.R
\name{effects2}
\alias{effects2}
\alias{effects2.lvmfit}
\alias{effects2.sCorrect}
\title{Effects From a Fitted Model After Small Sample Correction}
\usage{
effects2(object, ...)

\method{effects2}{lvmfit}(object, ssc = lava.options()$ssc,
  df = lava.options()$df, ...)

\method{effects2}{sCorrect}(object, link, conf.level = 0.95,
  robust = FALSE, ...)
}
\arguments{
\item{object}{an object that inherits from lvmfit.}

\item{...}{[logical] arguments passed to lower level methods.}

\item{ssc}{[logical] should the standard errors of the coefficients be corrected for small sample bias? Argument passed to \code{sCorrect}.}

\item{df}{[logical] should the degree of freedoms of the Wald statistic be computed using the Satterthwaite correction?
Otherwise the degree of freedoms are set to \code{Inf}, i.e. a normal distribution is used instead of a Student's t distribution when computing the p-values.}

\item{link}{[character vector] The path for which the effect should be assessed (e.g. \code{"A~B"}),
i.e. the effect of the right variable (B) on the left variable (A).}

\item{conf.level}{[numeric, 0-1] confidence level of the interval.}
}
\description{
Test whether a path in the latent variable model correspond to a null effect.
Similar to \code{lava::effects} but with small sample correction (if any).
So far it only work for paths composed of two edges.
}
\concept{small sample inference}
