% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate2.R
\name{estimate2}
\alias{estimate2}
\alias{.estimate2}
\title{Compute Bias Corrected Quantities.}
\usage{
.estimate2(
  object,
  epsilon,
  n.cluster,
  name.param,
  name.endogenous,
  name.meanparam,
  name.varparam,
  index.Omega,
  adjust.Omega,
  adjust.n,
  tol,
  n.iter,
  trace
)
}
\description{
Compute bias corrected residuals variance covariance matrix
and information matrix.
Also provides the leverage values and corrected sample size when adjust.n is set to TRUE.
}
\keyword{internal}
