% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-sscResiduals.R
\name{estimate2}
\alias{estimate2}
\alias{.sscResiduals}
\title{Compute Bias Corrected Quantities.}
\usage{
.sscResiduals(object, algorithm = "2")
}
\description{
Compute bias corrected residuals variance covariance matrix
and information matrix.
Also provides the leverage values and corrected sample size when adjust.n is set to TRUE.
}
\keyword{internal}
