% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-modelsearch2.R
\name{getStep}
\alias{getStep}
\alias{getStep.modelsearch2}
\title{Extract one Step From the Sequential Procedure}
\usage{
getStep(object, step, slot)

\method{getStep}{modelsearch2}(object, step = nStep(object), slot = NULL)
}
\arguments{
\item{object}{a \code{modelsearch2} object}

\item{step}{[integer >0] which test should be extracted?}

\item{slot}{[character] the element from the modelsearch2 object that should be extracted.}
}
\description{
Extract one step from the sequential procedure.
}
\examples{
\dontrun{
mSim <- lvm(Y~G+X1+X2)
addvar(mSim) <- ~Z1+Z2+Z3+Z4+Z5+Z6
df.data <- lava::sim(mSim, 1e2)

mBase <- lvm(Y~G)
addvar(mBase) <- ~X1+X2+Z1+Z2+Z3+Z4+Z5+Z6
e.lvm <- estimate(mBase, data = df.data)
res <- modelsearch2(e.lvm, method.p.adjust = "holm")

getStep(res)
getStep(res, slot = "sequenceTest")
getStep(res, step = 1)
}
}
\concept{extractor}
\concept{modelsearch}
