% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-Utils-nlme.R
\name{getVarCov2}
\alias{getVarCov2}
\alias{.getVarCov2}
\title{Reconstruct the marginal variance covariance matrix from a nlme model}
\usage{
.getVarCov2(object, ...)
}
\arguments{
\item{object}{a gls or lme object}

\item{...}{arguments to be passed to lower level methods.}
}
\description{
Reconstruct the marginal variance covariance matrix from a nlme model
}
\details{
The marginal variance covariance matrix for gls model is of the form:

\tabular{cccc}{
\eqn{\Sigma =} \tab \eqn{\sigma^2} \tab \eqn{\sigma^2 \sigma_2 \rho_{1,2}} \tab \eqn{\sigma^2 \sigma_3 \rho_{1,3}} \cr
\tab . \tab \eqn{\sigma^2 \sigma_2^2} \tab \eqn{\sigma^2 \sigma_3 \rho_{1,3}} \cr
\tab . \tab . \tab \eqn{\sigma^2 \sigma_3^2}
}

The marginal variance covariance matrix for lme model is of the form:
}
