% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlf2.R
\name{glht.mlf2}
\alias{glht.mlf2}
\title{General Linear Hypothesis}
\usage{
\method{glht}{mlf2}(model, linfct, ...)
}
\arguments{
\item{model}{a mmm2 object}

\item{linfct}{a specification of the linear hypotheses to be tested.
See the documentation of the \code{glht} function from the multcomp package.}

\item{...}{arguments passed to \code{glht}, \code{vcov}, and \code{lTest}.}
}
\description{
General linear hypotheses and multiple outcomes for mmm2 objects.
}
\examples{
## Simulate data
mSim <- lvm(c(Y1,Y2,Y3)~ beta * eta, E ~ 1)
latent(mSim) <- "eta"
set.seed(10)
n <- 1e2

df.data <- sim(mSim, n, latent = FALSE, p = c(beta = 1))

## Fit separate models
ls.lvm <- list(Y1 = estimate(lvm(Y1~E), data = dt.data),
               Y2 = estimate(lvm(Y2~E), data = dt.data),
               Y3 = estimate(lvm(Y3~E), data = dt.data))
class(ls.lvm) <- "ls.lvmfit"

## Create contrast matrix
C <- createContrast(ls.lvm, var.test = "E")

lvm.glht <- glht(ls.lvm, linfct = C)
summary(lvm.glht)
}
