% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-iid2.R
\name{iid2plot}
\alias{iid2plot}
\title{Display the i.i.d. Decomposition}
\usage{
iid2plot(object, param)
}
\arguments{
\item{object}{a \code{lvmfit} or \code{lvmfit2} object (i.e. output of \code{lava::estimate} or \code{lavaSearch2::estimate2}).}

\item{param}{[character] name of one of the model parameters.}
}
\description{
Extract the i.i.d. decomposition and display it along with the corresponding coefficient.
}
