% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lTest.R
\name{lTest}
\alias{lTest}
\alias{lTest.lm}
\alias{lTest.gls}
\alias{lTest.lme}
\alias{lTest.lvmfit}
\title{Test linear hypotheses using Wald statistic}
\usage{
lTest(object, ...)

\method{lTest}{lm}(object, C = NULL, ...)

\method{lTest}{gls}(object, C = NULL, ...)

\method{lTest}{lme}(object, C = NULL, ...)

\method{lTest}{lvmfit}(object, C = NULL, ...)
}
\arguments{
\item{object}{a \code{lvm} object.}

\item{...}{arguments to be passed to \code{compare2}.}

\item{C}{[optional] a contrast matrix.}
}
\description{
Test linear hypotheses using Wald statistic. Deprecated see compare2.
}
