% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsearch2.R
\name{modelsearch2}
\alias{modelsearch2}
\alias{modelsearch2.lvmfit}
\alias{modelsearch2.default}
\alias{.modelsearch2}
\title{Automatic extension of the lvm}
\usage{
modelsearch2(x, ...)

\method{modelsearch2}{lvmfit}(x, data = NULL, link = NULL,
  method.iid = "iidJack", exclude.var = NULL, rm.latent_latent = FALSE,
  rm.endo_endo = FALSE, rm.latent_endo = FALSE, ...)

\method{modelsearch2}{default}(x, link, data = NULL, method.iid = "iidJack",
  statistic = "Wald", ...)

.modelsearch2(x, link, restricted, directive, update.FCT, update.args, iid.FCT,
  packages = NULL, statistic = "Wald", conditional = FALSE,
  exposure = NULL, method.p.adjust = "max", nStep = NULL,
  na.omit = TRUE, alpha = 0.05, method.max = "integration", ncpus = 1,
  display.warnings = TRUE, trace = 1, export.iid = FALSE)
}
\arguments{
\item{x}{a lvm model}

\item{...}{additional arguments to be passed to lower levels functions.}

\item{data}{the dataset used to identify the model}

\item{link}{the links to consider when expanding the model}

\item{method.iid}{the method used to compute the influence function. Can be \code{"iid"} (first order approximation) or \code{"iidJack"} (jacknife estimate).}

\item{exclude.var}{see the documentation of \code{\link{findNewLink}}.}

\item{rm.latent_latent}{see the documentation of \code{\link{findNewLink}}.}

\item{rm.endo_endo}{see the documentation of \code{\link{findNewLink}}.}

\item{rm.latent_endo}{see the documentation of \code{\link{findNewLink}}.}

\item{statistic}{test statistic used to perform the test. Can the likelihood ratio test (LR), the score (score) or the max statistic (max).}

\item{restricted}{internal argument.}

\item{directive}{internal argument.}

\item{update.FCT}{internal argument.}

\item{update.args}{internal argument.}

\item{iid.FCT}{internal argument.}

\item{packages}{name of the R packages to be exported to each CPU.}

\item{conditional}{should the p.values be corrected for sequential testing. Experimental.}

\item{exposure}{the links to be tested at the end of the variable selection process. Not functional for now.}

\item{method.p.adjust}{the method used to adjust the p.values for multiple comparisons. Ignore when using the max statistic.}

\item{nStep}{the maximum number of links that can be added to the model.}

\item{na.omit}{should model leading to NA for the test statistic be ignored. Otherwise this will stop the selection process.}

\item{alpha}{the significance threshold for retaining a new link}

\item{method.max}{the method used to compute the distribution of the max statistic \code{integration}.}

\item{ncpus}{the number of cpus that can be used for the computations.}

\item{display.warnings}{should warnings be display? May occur when dealing with categorical variables or when fitting an extended model.}

\item{trace}{should the execution be traced?}

\item{export.iid}{should the iid decomposition of the retained coefficient be export. Only relevant when statistic is set to max.}
}
\value{
a latent variable model
}
\description{
Add all relevant path between the variables present in a lvm.
}
\examples{

# to save computation time 
method.iid <- "iid"

#### linear regression ####
mSim <- lvm(Y~X1+X2+X3+X4)
addvar(mSim) <- ~Z1+Z2
d <- sim(mSim,1e2)
eLM <- lm(Y~X1,data = d)

res <- modelsearch2(eLM, link = c("Y~X2","Y~X3","Y~X4","Y~Z1","Y~Z2"),
             statistic = "LR", method.p.adjust = "holm")
res <- modelsearch2(eLM, link = c("Y~X2","Y~X3","Y~X4","Y~Z1","Y~Z2"),
             statistic = "Wald", method.p.adjust = "holm",
             method.iid = method.iid)
res <- modelsearch2(eLM, link = c("Y~X2","Y~X3","Y~X4","Y~Z1","Y~Z2"),
                    method.iid = method.iid)

#### Cox model ####
library(survival)
data(Melanoma, package = "riskRegression")
m <- coxph(Surv(time,status==1)~ici+age, data = Melanoma, x = TRUE, y = TRUE)
res <- modelsearch2(m, link = c(status~epicel,status~sex),
             method.iid = method.iid, packages = "survival")
res

#### LVM ####
mSim <- lvm()
regression(mSim) <- c(y1,y2,y3)~u
regression(mSim) <- u~x1+x2
categorical(mSim,labels=c("A","B","C")) <- "x2"
latent(mSim) <- ~u
covariance(mSim) <- y1~y2
df <- sim(mSim, 1e2)
df$Id <- 1:NROW(df)

m <- lvm(c(y1,y2,y3)~u)
latent(m) <- ~u
addvar(m) <- ~x1+x2 

e <- estimate(m, df)

## score
\dontrun{
resScore <- modelsearch2(e, statistic = "score", method.p.adjust = "holm")
resLR <- modelsearch2(e, statistic = "LR", method.p.adjust = "holm")
resMax <- modelsearch2(e, rm.endo_endo = TRUE,
                        method.iid = method.iid, statistic = "Wald")
# resMax <- modelsearch2(e, rm.endo_endo = TRUE,
#                        method.iid = method.iid, statistic = "Wald", conditional = TRUE)
}
\dontshow{
links <- c(u~x1,u~x2C,y3~x2C)
resScore <- modelsearch2(e, statistic = "score", link = links, method.p.adjust = "holm")
resLR <- modelsearch2(e, statistic = "LR", link = links, method.p.adjust = "holm")
resMax <- modelsearch2(e, rm.endo_endo = TRUE, statistic = "Wald", method.iid = "iid", link = links)
}

}
