% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-modelsearchLR.R
\name{modelsearchLR}
\alias{modelsearchLR}
\title{Model searching using a likelihood ratio test}
\usage{
modelsearchLR(x, data, restricted, link, directive, update.FCT, update.args,
  method.p.adjust, display.warnings, trace)
}
\value{
a list
}
\description{
Model searching using a likelihood ratio test
}
\keyword{internal}
