% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsearchLR.R
\name{modelsearchLR}
\alias{modelsearchLR}
\title{Testing the Relevance of Additional Links Using a Likelihood Ratio Test}
\usage{
modelsearchLR(x, data, restricted, link, directive, update.FCT, update.args,
  method.p.adjust, display.warnings, trace)
}
\value{
a list
}
\description{
Testing the Relevance of Additional Links Using a Likelihood Ratio Test.
}
\keyword{internal}
