% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-modelsearchMax.R
\name{modelsearchMax}
\alias{modelsearchMax}
\title{Model searching using the max statistic}
\usage{
modelsearchMax(x, restricted, link, directive, packages, update.FCT,
  update.args, iid.FCT, alpha, method.p.adjust, method.max, n.sim = 1000,
  iid.previous = NULL, quantile.previous = NULL, export.iid, trace, ncpus,
  initCpus)
}
\value{
an object of class lvmfit
}
\description{
Model searching using the max statistic to retain or not a link
}
\seealso{
\code{link{modelsearch2}}
}
\keyword{internal}
