% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsearchMax.R
\name{modelsearchMax}
\alias{modelsearchMax}
\title{Model searching using the max statistic}
\usage{
modelsearchMax(x, data = NULL, restricted, link, directive,
  method.max = "boot-wild", alpha = 0.05, ls.LVMargs = NULL,
  method.p.adjust = "bonferroni", n.sim = 10000, conditional = NULL,
  mu.conditional = NULL, iid.conditional = NULL, export.iid = 1,
  trace = 1, ncpus = 1)
}
\value{
an object of class lvmfit
}
\description{
Model searching using the max statistic to retain or not a link
}
\seealso{
\code{link{modelsearch2}}
}
\keyword{internal}
