% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsearchMax.R
\name{modelsearchMax}
\alias{modelsearchMax}
\title{Testing the Relevance of Additional Links Using the Max Statistic}
\usage{
modelsearchMax(x, restricted, link, directive, packages, update.FCT,
  update.args, iid.FCT, alpha, method.p.adjust, method.max, n.sim = 1000,
  iid.previous = NULL, quantile.previous = NULL, export.iid, trace, ncpus,
  init.cpus)
}
\value{
A \code{lvmfit} object.
}
\description{
Testing the Relevance of Additional Links Using the Max Statistic.
}
\seealso{
\code{link{modelsearch2}}
}
\concept{
modelsearch
}
\keyword{internal}
