% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residplot.R
\name{residplot}
\alias{residplot}
\alias{residplot.lvmfit}
\title{Plot the Residuals of a LVM Object}
\usage{
residplot(object, ...)

\method{residplot}{lvmfit}(object, res.variables = endogenous(object),
  obs.variables = res.variables, sd.kernel = 0.5, kernel = "dnorm",
  plot.weights = FALSE, ncol = NULL, smooth.mean = TRUE,
  smooth.sd = TRUE, plot = TRUE, ...)
}
\arguments{
\item{object}{a lvm model.}

\item{...}{additional arguments.}

\item{res.variables}{the endogenous variable for which the residuals should be displayed.}

\item{obs.variables}{same as res.variables or a variable present in the model}

\item{sd.kernel}{the standard deviation of the kernel used to smooth the variance.}

\item{kernel}{the type of kernel used to smooth the variance.}

\item{plot.weights}{should the weights used to compute the variance be displayed?}

\item{ncol}{the number of columns in the graphical display.}

\item{smooth.mean}{should the mean be displayed across the obs.variables values?}

\item{smooth.sd}{should the variance be displayed across the obs.variables values?}

\item{plot}{should the fit be displayed in a graphical window.}
}
\value{
a list containing:
\itemize{
\item plot the display of the fit
\item data the dataset used to make the display
}
}
\description{
Plot the residuals for each outcome of a lvm object.
}
\examples{
m <- lvm(y1~eta+x1,y2~eta,y3~eta+x3)
latent(m) <- ~ eta
dd <- lava::sim(m,100) ## Simulate 100 observations from model
e <- estimate(m, dd) ## Estimate parameters

res <- residplot(e)
residplot(e, obs.variables = "x1")

m <- lvm(y~x)
distribution(m,~y) <- function(n,mean,x) rnorm(n,mean,exp(x)^.5)
d <- lava::sim(m,1e3)

e <- estimate(m,data = d)

residplot(e)
# residplot(e, plot.weights = TRUE)
residplot(e, res.variables = "y", obs.variables = "x")
}
