% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-residuals2.R
\name{residuals2}
\alias{residuals2}
\alias{residuals2.sCorrect}
\title{Extract Residuals After Small Sample Correction.}
\usage{
residuals2(object, type, format)

\method{residuals2}{sCorrect}(object, type = "response", format = "wide")
}
\arguments{
\item{object}{a \code{lm2}, \code{gls2}, \code{lme2}, or \code{lvmfit2} object.}

\item{type}{[character] the type of residual to extract:
\code{"response"} for raw residuals,
\code{"studentized"} for studentized residuals,
\code{"normalized"} for normalized residuals.}

\item{format}{[character] Use \code{"wide"} to return the residuals in the wide format (one row relative to each sample).
Otherwise use \code{"long"} to return the residuals in the long format.}
}
\value{
a matrix containing the residuals relative to each sample (in rows)
and each endogenous variable (in column).
}
\description{
Extract residuals from \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} objects.
Similar to \code{residuals} but with small sample correction.
}
\details{
If argument \code{p} or \code{data} is not null, then the small sample size correction is recomputed to correct the residuals. \cr

The raw residuals are defined by  observation minus the fitted value:
\deqn{
\varepsilon = (Y_1 - \mu_1, ..., Y_m - \mu_m)
}
The studentized residuals divided the raw residuals relative to each endogenous variable by the modeled variance of the endogenous variable.
\deqn{
\varepsilon_{stud} =(\frac{Y_1 - \mu_1}{\sigma_1}, ..., \frac{Y_m - \mu_m}{\sigma_m})
}
The normalized residuals multiply the raw residuals by the inverse of the square root of the modeled residual variance covariance matrix.
\deqn{
\varepsilon_{norm} = \varepsilon \Omega^{-1/2}
}
}
\examples{
set.seed(10)
n <- 101

#### linear regression ####
Y1 <- rnorm(n, mean = 0)
Y2 <- rnorm(n, mean = 0.3)
Id <- findInterval(runif(n), seq(0.1,1,0.1))
data.df <- rbind(data.frame(Y=Y1,G="1",Id = Id),
           data.frame(Y=Y2,G="2",Id = Id)
           )
m.lm <- lm(Y ~ G, data = data.df)
residuals2(m.lm)

#### mixed models ####
library(nlme)
m.gls <- gls(Y ~ G, weights = varIdent(form = ~ 1|Id), data = data.df)
coef2(m.gls)
m.lme <- lme(Y ~ G, random = ~ 1|Id, data = data.df)
coef2(m.lme)

#### latent variable models ####
library(lava)
e.lvm <- estimate(lvm(Y ~ G), data = data.df)
coef2(e.lvm)

## simulate data
set.seed(10)
m <- lvm(Y1~eta,Y2~eta,Y3~eta)
latent(m) <- ~eta
d <- lava::sim(m,20, latent = FALSE)

## standard linear model
e.lm <- lm(Y1~Y2, data = d)
sCorrect(e.lm) <- TRUE

sigma(e.lm)^2
mean(residuals(e.lm)^2)
mean(residuals2(e.lm)^2)

## latent variable model
e.lvm <- estimate(m, data = d)
sCorrect(e.lvm) <- TRUE
mean(residuals2(e.lvm)^2)

}
\seealso{
\code{\link{sCorrect}} to obtain \code{lm2}, \code{gls2}, \code{lme2}, or \code{lvmfit2} objects.
}
\concept{small sample inference}
