% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score2.R
\name{score2-internal}
\alias{score2-internal}
\alias{.score2}
\title{Compute the Corrected Score.}
\usage{
.score2(
  epsilon,
  Omega,
  OmegaM1,
  dmu,
  dOmega,
  name.param,
  name.meanparam,
  name.varparam,
  index.Omega,
  n.cluster,
  indiv
)
}
\arguments{
\item{n.cluster}{[integer >0] the number of observations.}
}
\description{
Compute the corrected score when there is no missing value.
}
\keyword{internal}
