% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect-score2.R
\name{score2}
\alias{score2}
\alias{score2.sCorrect}
\title{Extract the Individual Score.}
\usage{
score2(object, indiv, cluster, as.lava)

\method{score2}{sCorrect}(object, indiv = FALSE, cluster = NULL, as.lava = TRUE)
}
\arguments{
\item{object}{a \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} object.}

\item{indiv}{[logical] If \code{TRUE}, the score relative to each observation is returned.
Otherwise the total score is returned.}

\item{cluster}{[integer vector] the grouping variable relative to which the observations are iid.}

\item{as.lava}{[logical] Should the order and the name of the coefficient be the same as those obtained using coef with type = -1.
Only relevant for \code{lvmfit} objects.}
}
\value{
A matrix containing the score relative to each sample (in rows)
and each model coefficient (in columns).
}
\description{
Extract the individual score from \code{lm}, \code{gls}, \code{lme}, or \code{lvmfit} objects.
Similar to \code{lava::score} but compatible with small sample correction.
}
\details{
If argument \code{p} or \code{data} is not null, then the small sample size correction is recomputed to correct the influence function.
}
\examples{
n <- 5e1
p <- 3
X.name <- paste0("X",1:p)
link.lvm <- paste0("Y~",X.name)
formula.lvm <- as.formula(paste0("Y~",paste0(X.name,collapse="+")))

m <- lvm(formula.lvm)
distribution(m,~Id) <- sequence.lvm(0)
set.seed(10)
d <- lava::sim(m,n)

## linear model
e.lm <- lm(formula.lvm,data=d)
eS.lm <- sCorrect(e.lm, ssc = NA, df = NA)
score.tempo <- score2(eS.lm, indiv = TRUE)
colMeans(score.tempo)

## latent variable model
e.lvm <- estimate(lvm(formula.lvm),data=d)
eS.lvm <- sCorrect(e.lvm, df = NA, ssc = NA)
score.tempo <- score2(eS.lvm, indiv = TRUE)
range(score.tempo-score(e.lvm, indiv = TRUE))

}
\seealso{
\code{\link{sCorrect}} to obtain \code{lm2}, \code{gls2}, \code{lme2}, or \code{lvmfit2} objects.
}
\concept{small sample inference}
