% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect.R
\name{sCorrect}
\alias{sCorrect}
\alias{sCorrect.lm}
\alias{sCorrect.gls}
\alias{sCorrect.lme}
\alias{sCorrect.lvmfit}
\alias{sCorrect.lvmfit2}
\alias{sCorrect<-}
\alias{sCorrect<-.lm}
\alias{sCorrect<-.gls}
\alias{sCorrect<-.lme}
\alias{sCorrect<-.lvmfit}
\alias{sCorrect<-.lvmfit2}
\title{Compute the Derivative of the Information Matrix}
\usage{
sCorrect(object, ...)

\method{sCorrect}{lm}(object, bias.correct = FALSE, return.score = FALSE,
  ...)

\method{sCorrect}{gls}(object, cluster, vcov.param = NULL,
  bias.correct = FALSE, numericDerivative = FALSE, return.score = FALSE,
  ...)

\method{sCorrect}{lme}(object, cluster, vcov.param = NULL,
  bias.correct = FALSE, numericDerivative = FALSE, return.score = FALSE,
  ...)

\method{sCorrect}{lvmfit}(object, vcov.param = NULL, bias.correct = TRUE,
  numericDerivative = FALSE, return.score = FALSE, ...)

\method{sCorrect}{lvmfit2}(object, ...)

sCorrect(x, ...) <- value

\method{sCorrect}{lm}(x, ...) <- value

\method{sCorrect}{gls}(x, ...) <- value

\method{sCorrect}{lme}(x, ...) <- value

\method{sCorrect}{lvmfit}(x, ...) <- value

\method{sCorrect}{lvmfit2}(x, ...) <- value
}
\arguments{
\item{object}{a gls, lme, or lvm object.}

\item{...}{[internal] Only used by the generic method.}

\item{bias.correct}{[logical] should a small sample correction for the standard errors of the coefficients be performed ?}

\item{return.score}{[for internal use] export the score.}

\item{cluster}{the grouping variable relative to which the observations are iid.
Only required for gls models with no correlation argument.}

\item{vcov.param}{the variance-covariance matrix of the estimates.}

\item{numericDerivative}{If TRUE, the degree of freedom are computed using a numerical derivative.}

\item{x}{same as object.}

\item{value}{same as bias.correct.}
}
\description{
Compute the derivative of the information matrix.
}
