% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCorrect.R
\name{sCorrect}
\alias{sCorrect}
\alias{sCorrect.default}
\alias{sCorrect<-}
\alias{sCorrect<-.default}
\title{Depreciated Method For Small Sample Correction}
\usage{
sCorrect(object, ...)

\method{sCorrect}{default}(object, ...)

sCorrect(x, ...) <- value

\method{sCorrect}{default}(x, ...) <- value
}
\arguments{
\item{object, x}{a \code{lvmfit} object.}

\item{...}{not used.}

\item{value}{not used.}
}
\description{
Depreciated method for small sample correction, now replaced by the \code{\link{estimate2}} method.
}
