% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils-formula.R
\name{selectRegressor}
\alias{selectRegressor}
\alias{selectRegressor.formula}
\title{Regressor of a Formula.}
\usage{
selectRegressor(object, ...)

\method{selectRegressor}{formula}(object, format = "call", ...)
}
\arguments{
\item{object}{a formula}

\item{...}{[internal] Only used by the generic method.}

\item{format}{[character] should an object of format call be returned (\code{format = "call"}),
or the names of the variables (\code{format = "vars"})}
}
\description{
Return the regressor variables contained in the formula
}
\examples{

\dontrun{

selectRegressor <- lavaSearch2:::selectRegressor
selectRegressor.formula <- lavaSearch2:::selectRegressor.formula

selectRegressor(Y1~X1+X2)
selectRegressor(Y1~X1+X2, format = "vars")

selectRegressor(Y1~X1+Y1)
selectRegressor(Y1+Y2~X1+Y1, format = "vars")

selectRegressor(~X1+X2)
selectRegressor(~X1+X2, format = "vars")

}
}
\keyword{internal}
