% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.calibrateType1.R
\name{summary.calibrateType1}
\alias{summary.calibrateType1}
\title{Display the Type 1 Error Rate}
\usage{
\method{summary}{calibrateType1}(
  object,
  robust = FALSE,
  type = "type1error",
  alpha = 0.05,
  log.transform = TRUE,
  digits = 5,
  print = TRUE,
  ...
)
}
\arguments{
\item{object}{output of the \code{calibrateType1} function.}

\item{robust}{[character] should the results be displayed for both model-based and robust standard errors (\code{TRUE}),
only  model-based standard error (\code{FALSE}), or only robust standard error (\code{"only"})?}

\item{type}{[character] should the type 1 error rate be diplayed (\code{"type1error"}) or the bias (\code{"bias")}.}

\item{alpha}{[numeric, 0-1] the confidence levels.}

\item{log.transform}{[logical] should the confidence intervals be computed on the logit scale.}

\item{digits}{[integer >0] the number of decimal places to use when displaying the summary.}

\item{print}{should the summary be printed in the terminal.}

\item{...}{[internal] only used by the generic method.}
}
\description{
Display the type 1 error rate from the simulation results.
}
